### 2024-03-10 #6459 @editorialbot
[PRE REVIEW]: cuallee: A python package for data quality checks across multiple DataFrame APIs
**Submitting author:** @canimus (<a href="http://orcid.org/0000-0003-1937-8006">Herminio Vazquez</a>)
**Repository:** https://github.com/canimus/cuallee
**Branch with paper.md** (empty if default branch): main
**Version:** v0.8.8
