### 2024-03-10 109b5eb
Remove unnecessary figures
