+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "cuallee"
-version = "0.8.8"
+version = "0.9.0"
 authors = [
   { name="Herminio Vazquez", email="canimus@gmail.com"},
   { name="Virginie Grosboillot", email="vestalisvirginis@gmail.com" }
@@ -58,6 +58,9 @@ dagster = [
 cloud = [
   "msgpack == 1.0.8",
   "requests>=2.28",
+]
+pdf = [
+  "fpdf2==2.7.8"
 ]
 
 [tool.ruff.lint]
