+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
 name = cuallee
-version = 0.8.8
+version = 0.9.0
 [options]
 packages = find:
