### 2024-03-17 5835e22
Added information gain
