### 2024-03-17 3097c56
Added infogain check across dataframes
