### 2024-03-17 4e5d494
Added info gain test cases for all frames
