### 2024-03-17 7053c05
Added intelligence control
