+++ cuallee/polars_validation.py
@@ -148,7 +148,7 @@ class Compute:
         return Compute._result(
             dataframe.select(pl.col(rule.column).n_unique() == rule.value).to_series()
         )
-    
+
     def has_infogain(self, rule: Rule, dataframe: pl.DataFrame) -> Union[bool, int]:
         """More than 1 value in column"""
         return Compute._result(
