+++ test/unit/class_control/test_methods.py
@@ -10,19 +10,13 @@ def test_has_information():
     assert hasattr(Control, "information")
 
 
+def test_has_information():
+    assert hasattr(Control, "intelligence")
+
+
 def test_completeness_result():
     df = pd.DataFrame({"A": [1, 2, 3, 4, 5]})
     assert Control.completeness(df).status.eq("PASS").all()
-
-
-def test_information_result():
-    df = pd.DataFrame({"A": [1, 2, 3, 4, 5]})
-    assert Control.information(df).status.eq("PASS").all()
-
-
-def test_no_information_result():
-    df = pd.DataFrame({"A": [1, 1, 1, 1, 1]})
-    assert Control.information(df).status.eq("FAIL").any()
 
 
 def test_fillness(spark):
@@ -33,3 +27,8 @@ def test_fillness(spark):
 def test_emptyness(spark):
     df = spark.range(10)
     assert Control.percentage_empty(df) == 0
+
+
+def test_intelligence_result(spark):
+    df = spark.range(10)
+    assert Control.information(df) == ["id"]
