### 2024-03-17 9a71181
Feature intelligence (#179)
* Added report and version bump

* Added information gain

* Added infogain check across dataframes

* Added info gain test cases for all frames

* Added infogain tests

* Added intelligence control

* Added format

* Working inforrmation test
