+++ cuallee/__init__.py
@@ -9,9 +9,10 @@ from datetime import datetime, timedelta, timezone
 from types import ModuleType
 from typing import Any, Dict, List, Literal, Optional, Protocol, Tuple, Union
 from toolz import compose, valfilter  # type: ignore
+from toolz.curried import map as map_curried
 
 logger = logging.getLogger("cuallee")
-__version__ = "0.8.8"
+__version__ = "0.9.0"
 # Verify Libraries Available
 # ==========================
 try:
@@ -43,7 +44,6 @@ try:
     from google.cloud import bigquery
 except (ModuleNotFoundError, ImportError):
     logger.debug("KO: BigQuery")
-
 
 
 class CheckLevel(enum.Enum):
@@ -185,12 +185,12 @@ class ComputeEngine(Protocol):
 class Check:
     def __init__(
         self,
-        level: Union[CheckLevel, int],
-        name: str,
+        level: Union[CheckLevel, int] = 0,
+        name: str = "cuallee.check",
         *,
         execution_date: datetime = datetime.now(timezone.utc),
         table_name: str = None,
-        session: Any = None
+        session: Any = None,
     ):
         """A container of data quality rules."""
         self._rule: Dict[str, Rule] = {}
@@ -209,6 +209,7 @@ class Check:
         self.table_name = table_name
         try:
             from .iso.checks import ISO
+
             self.iso = ISO(self)
         except (ModuleNotFoundError, ImportError):
             logger.error("ISO module requires requests")
@@ -552,6 +553,20 @@ class Check:
         Rule("has_cardinality", column, value, CheckDataType.AGNOSTIC) >> self._rule
         return self
 
+    def has_infogain(self, column: str, pct: float = 1.0):
+        """Validate cardinality > 1"""
+        (
+            Rule(
+                method="has_infogain",
+                column=column,
+                value="N/A",
+                data_type=CheckDataType.AGNOSTIC,
+                coverage=pct,
+            )
+            >> self._rule
+        )
+        return self
+
     def has_entropy(self, column: str, value: float, tolerance: float = 0.01):
         """Validation for entropy calculation on continuous values"""
         (
@@ -705,14 +720,23 @@ class Control:
         return check.validate(dataframe)
 
     @staticmethod
+    def information(dataframe, **kwargs):
+        """Information gain"""
+        check = Check(CheckLevel.WARNING, name="Information", **kwargs)
+        [check.is_complete(c) for c in dataframe.columns]
+        [check.has_infogain(c) for c in dataframe.columns]
+        [check.is_legit(c) for c in dataframe.columns]
+        return check.validate(dataframe)
+
+    @staticmethod
     def percentage_fill(dataframe, **kwargs):
         """Control the percentage of values filled"""
-        from toolz.curried import map as map_curried
+
         compute = compose(
             map_curried(operator.attrgetter("pass_rate")),
             operator.methodcaller("collect"),
-            operator.methodcaller("select", "pass_rate")
-            )
+            operator.methodcaller("select", "pass_rate"),
+        )
         result = list(compute(Control.completeness(dataframe, **kwargs)))
         return sum(result) / len(result)
 
@@ -720,3 +744,19 @@ class Control:
     def percentage_empty(dataframe, **kwargs):
         """Control the percentage of values empty"""
         return 1 - Control.percentage_fill(dataframe, **kwargs)
+
+    @staticmethod
+    def intelligence(dataframe, **kwargs) -> List[str]:
+        """Return worthy columns"""
+        complete_gain = compose(
+            list,
+            map_curried(lambda x: x.column),
+            operator.methodcaller("collect"),
+            operator.methodcaller("distinct"),
+            operator.methodcaller("select", "column"),
[…]
