### 2024-03-17 7166eaa
Feature fix bigdf fail (#180)
* Removed extra requests

* Added recognized font

* Fixed the PASS ratio issue on spark
