+++ cuallee/report/__init__.py
@@ -5,10 +5,10 @@ from datetime import datetime, timezone
 
 def pdf(data: List[Tuple[str]], name: str = "cuallee.pdf"):
     today = datetime.now(timezone.utc).strftime("%Y%m%d%H%M%S")
-    style = FontFace(fill_color=rgb_7)
+    style = FontFace(fill_color="#AAAAAA")
     pdf = FPDF(orientation="landscape", format="A4")
     pdf.add_page()
-    pdf.set_font("Berkeley", size=6)
+    pdf.set_font("Helvetica", size=6)
     irow = -1
     with pdf.table(
         borders_layout="SINGLE_TOP_LINE",
@@ -18,9 +18,7 @@ def pdf(data: List[Tuple[str]], name: str = "cuallee.pdf"):
         for data_row in data:
             row = table.row()
             for datum in data_row:
-                # if irow in top_5:
-                #     row.cell(datum, style=style)
-                # else:
-                #    row.cell(datum)
                 row.cell(datum)
             irow += 1
+
+    pdf.output(name)
