+++ pyproject.toml
@@ -57,7 +57,6 @@ dagster = [
 ]
 cloud = [
   "msgpack == 1.0.8",
-  "requests>=2.28",
 ]
 pdf = [
   "fpdf2==2.7.8"
