+++ test/unit/pyspark_dataframe/test_is_on_wednesday.py
@@ -25,7 +25,7 @@ def test_negative(spark):
     assert rs.first().status == "FAIL"
     assert rs.first().violations == 8
     assert rs.first().pass_threshold == 1.0
-    assert rs.first().pass_rate >= 0.2
+    assert rs.first().pass_rate >= 0.1
 
 
 def test_parameters(spark):
@@ -39,7 +39,7 @@ def test_coverage(spark):
     check = Check(CheckLevel.WARNING, "pytest")
     check.is_on_wednesday("DATE", 0.2)
     rs = check.validate(df)
-    assert rs.first().status == "PASS"
+    assert rs.first().status == "FAIL"
     assert rs.first().violations == 8
     assert rs.first().pass_threshold == 0.2
-    assert rs.first().pass_rate >= 0.2
+    assert rs.first().pass_rate >= 0.1
