### 2024-03-22 663b914
Added spark connect implementation
