### 2024-03-22 18a2e9c
Added spark connect implementation (#185)
