+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
 name = cuallee
-version = 0.9.0
+version = 0.9.1
 [options]
 packages = find:
