### 2024-03-22 5ad2580
modified a deprecated horizontal sum operation
