+++ pyproject.toml
@@ -24,7 +24,7 @@ dependencies = [
 
 [project.optional-dependencies]
 dev = [
-  "black==24.2.0",
+  "black==24.3.0",
   "ruff==0.3.3"
 ]
 pyspark = [
