### 2024-03-23 8076747
Bump dagster from 1.6.9 to 1.6.11 (#181)
Bumps [dagster](https://github.com/dagster-io/dagster) from 1.6.9 to 1.6.11.
- [Release notes](https://github.com/dagster-io/dagster/releases)
- [Changelog](https://github.com/dagster-io/dagster/blob/master/CHANGES.md)
- [Commits](https://github.com/dagster-io/dagster/compare/1.6.9...1.6.11)

---
updated-dependencies:
- dependency-name: dagster
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.git
[…]
