+++ test/unit/pyspark_dataframe/test_spark_validation.py
@@ -240,6 +240,7 @@ minversion = pytest.mark.skipif(
 )
 
 
+# FIX THIS
 @minversion
 def test_observe_method_return_tuple(spark):
     df = spark.range(10)
@@ -254,7 +255,7 @@ def test_observe_method_return_tuple(spark):
     assert row == df.count()
     assert len(observe) == 1
 
-
+# FIX THIS
 def test_observe_no_compute(spark):
     df = spark.range(10)
     rs = (
@@ -280,14 +281,14 @@ def test_compute_transform_method(spark):
     assert isinstance(transform, Dict)
     assert len(transform) == 1
 
-
+# FIX THIS
 def test_compute_summary_return_dataframe(spark):
     df = spark.range(10)
     check = Check(CheckLevel.WARNING, "test_spark_dataframe").is_complete("id")
     rs = PSV.summary(check, df)
     assert isinstance(rs, DataFrame)
 
-
+# FIX THIS
 @patch.object(SparkSession, "version", "3.2.0")
 def test_lower_spark_version(spark):
     df = spark.range(10)
