+++ conftest.py
@@ -23,12 +23,18 @@ except:
 @pytest.fixture(scope="session")
 def spark():
     try:
-        os.environ["SPARK_REMOTE"] = "sc://localhost:15002"
         logger = logging.getLogger("py4j")
         logger.setLevel(logging.ERROR)
-        spark_session = SparkSession.builder.config(
-            "spark.driver.memory", "2g"
-        ).getOrCreate()
+
+        if "CUALLEE_SPARK_REMOTE" in os.environ or "SPARK_REMOTE" in os.environ:
+            spark_remote_url = os.environ.get("CUALLEE_SPARK_REMOTE") or os.environ.get("SPARK_REMOTE")
+            spark_session = SparkSession.builder.remote(spark_remote_url).getOrCreate()
+
+        else:
+            spark_session = SparkSession.builder.config(
+                "spark.driver.memory", "2g"
+            ).getOrCreate()
+
         yield spark_session
     except:
         pass
@@ -86,22 +92,24 @@ def db() -> duckdb.DuckDBPyConnection:
 
 @pytest.fixture(scope="session")
 def bq_client():
-
-    if Path('temp/key.json').exists()==True:
-        credentials = os.getenv('GOOGLE_APPLICATION_CREDENTIALS')
+    if Path("temp/key.json").exists() == True:
+        credentials = os.getenv("GOOGLE_APPLICATION_CREDENTIALS")
     else:
         from google.oauth2 import service_account
         import os
         import json
+
         with open("key.json", "w") as writer:
             json.dump(json.loads(os.getenv("GOOGLE_KEY")), writer)
 
         credentials = service_account.Credentials.from_service_account_file("key.json")
 
     try:
-        client = bigquery.Client(project="cuallee-bigquery-386709", credentials=credentials)
+        client = bigquery.Client(
+            project="cuallee-bigquery-386709", credentials=credentials
+        )
         return client
     except:
         pass
-    #finally:
-        #client.stop()
+    # finally:
+    # client.stop()
