+++ pyproject.toml
@@ -45,7 +45,7 @@ bigquery =[
   "pyarrow >= 11.0.0"
 ]
 duckdb = [
-  "duckdb==0.10.0"
+  "duckdb==0.10.1"
 ]
 polars = [
   "polars>=0.19.6"
