### 2024-03-25 f39682f
- Added `is_complete`, `are_complete`
