+++ pyproject.toml
@@ -64,7 +64,7 @@ cloud = [
 pdf = [
   "fpdf2==2.7.8"
 ]
-getdaft = [
+daft = [
   "getdaft == 0.2.19"
 ]
 
