### 2024-03-25 c1d2eba
- Added `is_unique` & `are_unique`
