### 2024-03-25 174ba81
- Added `is_greater_than`
