+++ cuallee/daft_validation.py
@@ -31,8 +31,8 @@ class Compute:
         return dataframe.select(*perdicate).to_pandas().nunique().sum() / len(rule.column)
 
     def is_greater_than(self, rule: Rule, dataframe: daft.DataFrame) -> int:
-        perdicate = (daft.col(rule.column) > rule.value).cast(daft.DataType.int64())
-        return dataframe.select(perdicate).sum(perdicate).to_pandas().iloc[0, 0]
+        perdicate = (daft.col(rule.column) > rule.value).cast(daft.DataType.int64()).sum()
+        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
 
 
 def compute(rules: Dict[str, Rule]):
