+++ test/unit/daft copy/test_is_greater_than.py
@@ -1,6 +1,7 @@
 import daft
 import pytest
 import numpy as np
+import pandas as pd
 
 from cuallee import Check
 

[…]
