### 2024-03-26 c8777c4
- Added `has_pattern`
