+++ cuallee/daft_validation.py
@@ -50,6 +50,12 @@ class Compute:
         perdicate = (daft.col(rule.column) == rule.value).cast(daft.DataType.int64()).sum()
         return dataframe.select(perdicate).to_pandas().iloc[0, 0]
 
+    def has_pattern(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        perdicate = (daft.col(rule.column).str.match(rule.value)).cast(daft.DataType.int64()).sum()
+        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+
+
+
 def compute(rules: Dict[str, Rule]):
     """Daft computes directly on the predicates"""
     return True
