### 2024-03-26 713702e
- Added `has_min`, `has_max`, `has_std`, `has_mean`, `has_sum`
