### 2024-03-26 0ad39f4
- Added `has_cardinality` and `has_infogain`
