### 2024-03-26 1f0169e
- Added `has_max_by` and `has_min_by`
