### 2024-03-26 835e67a
- Added `has_correlation`
