### 2024-03-26 7479899
- Added `is_on_weekday` , `is_on_weekend` , `is_on_monday` , `is_on_tuesday` , `is_on_wednesday` , `is_on_thursday` , `is_on_friday`, `is_on_saturday`, `is_on_sunday`
