### 2024-03-26 141d63e
- Added `is_inside_interquartile_range`
