### 2024-03-27 4e9c851
- Added `is_daily`
