### 2024-03-27 29e9b9a
- Updated `has_std`
