### 2024-03-27 47b5d39
Added `has_entropy`
