### 2024-03-27 852da38
- Added test for `test_positive`, `is_negative`, `is_in_millions`, `is_in_billions`
