### 2024-03-27 924b259
- Added tests to `test_validate.py`
