+++ test/unit/daft/test_has_infogain.py
@@ -1,5 +1,4 @@
 import daft
-import pytest
 import numpy as np
 
 from cuallee import Check
