+++ test/unit/daft/test_is_negative.py
@@ -1,6 +1,5 @@
 import daft
 import pytest
-import numpy as np
 
 from cuallee import Check
 
