### 2024-03-27 bf15a64
- Added `satisfies`
