### 2024-03-27 1cb909c
Daft Integration (#192)
* - Added `is_complete`, `are_complete`

* - Fixed pyproject.toml

* - Added `is_unique` & `are_unique`

* - Added `is_greater_than`

* - Added: `is_greater_or_equal_than` , `is_less_than` , `is_less_or_equal_than` , `is_equal_than`

* - Added `has_pattern`

* - Added `has_min`, `has_max`, `has_std`, `has_mean`, `has_sum`

* - Added `has_cardinality` and `has_infogain`

* - Added `is_between`

* - Added `is_contained_in` and `not_contained_in`

* - Added `has_percentile`

* - Added `has_max_by` and `has_min_by`

* - Added `has_correlation`

* - Added `is_on_weekday` , `
[…]
