+++ cuallee/daft_validation.py
@@ -1,0 +1,370 @@
+import daft
+import operator
+import statistics
+import numpy as np
+import pandas as pd
+
+from math import log2
+from toolz import first
+from typing import Union
+from numbers import Number
+from typing import Dict, List
+
+from itertools import groupby
+from operator import itemgetter
+
+from cuallee import Check, Rule
+
+
+class Compute:
+
+    def is_complete(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        perdicate = daft.col(rule.column).not_null().cast(daft.DataType.int64()).sum()
+        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+
+    def are_complete(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        col_names = rule.column
+        perdicate = [ daft.col(col_name).not_null().cast(daft.DataType.int64()).sum() for col_name in col_names]
+        return dataframe.select(*perdicate).to_pandas().astype(int).sum().sum() / len(col_names)
+
+    def is_unique(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        perdicate = daft.col(rule.column)
+        return dataframe.select(perdicate).distinct().count(perdicate).to_pandas().iloc[0, 0]
+
+    def are_unique(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        # TODO: Find a way to do this in daft and not pandas
+        perdicate = [ daft.col(col_name) for col_name in rule.column]
+        return dataframe.select(*perdicate).to_pandas().nunique().sum() / len(rule.column)
+
+    def is_greater_than(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        perdicate = (daft.col(rule.column) > rule.value).cast(daft.DataType.int64()).sum()
+        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+
+    def is_greater_or_equal_than(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        perdicate = (daft.col(rule.column) >= rule.value).cast(daft.DataType.int64()).sum()
+        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+
+    def is_less_than(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        perdicate = (daft.col(rule.column) < rule.value).cast(daft.DataType.int64()).sum()
+        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+
+    def is_less_or_equal_than(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        perdicate = (daft.col(rule.column) <= rule.value).cast(daft.DataType.int64()).sum()
+        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+
+    def is_equal_than(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        perdicate = (daft.col(rule.column) == rule.value).cast(daft.DataType.int64()).sum()
+        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+
+    def has_pattern(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        perdicate = (daft.col(rule.column).str.match(rule.value)).cast(daft.DataType.int64()).sum()
+        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+
+    def has_min(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        perdicate = daft.col(rule.column).min()
[…]
