+++ test/unit/daft/test_is_on_saturday.py
@@ -4,6 +4,7 @@ import pendulum as lu
 
 from cuallee import Check
 
+
 def test_positive(check: Check):
     check.is_on_saturday("id")
     pd_df = pd.DataFrame(
@@ -35,4 +36,6 @@ def test_coverage(check: Check):
     df = daft.from_pandas(pd_df)
     result = check.validate(df)
     assert result.select(daft.col("status").str.match("PASS")).to_pandas().status.all()
-    assert result.select(daft.col("pass_rate").max() == 1 / 7).to_pandas().pass_rate.all()
+    assert (
+        result.select(daft.col("pass_rate").max() == 1 / 7).to_pandas().pass_rate.all()
+    )
