+++ test/unit/daft/test_are_complete.py
@@ -3,6 +3,7 @@ import pytest
 
 from cuallee import Check
 
+
 def test_positive(check: Check):
     check.are_complete(("id", "id2"))
     df = daft.from_pydict({"id": [10, 20], "id2": [300, 500]})
@@ -31,4 +32,6 @@ def test_coverage(check: Check):
     check.are_complete(("id", "id2"), 0.75)
     df = daft.from_pydict({"id": [10, None], "id2": [300, 500]})
     result = check.validate(df)
-    assert result.select(daft.col("pass_rate").max() == 0.75).to_pandas().pass_rate.all()
+    assert (
+        result.select(daft.col("pass_rate").max() == 0.75).to_pandas().pass_rate.all()
+    )
