+++ test/unit/daft/test_is_on_weekend.py
@@ -36,4 +36,6 @@ def test_coverage(check: Check):
     df = daft.from_pandas(pd_df)
     result = check.validate(df)
     assert result.select(daft.col("status").str.match("PASS")).to_pandas().status.all()
-    assert result.select(daft.col("pass_rate").max() == 2 / 7).to_pandas().pass_rate.all()
+    assert (
+        result.select(daft.col("pass_rate").max() == 2 / 7).to_pandas().pass_rate.all()
+    )
