+++ test/unit/daft/test_is_between.py
@@ -23,8 +23,13 @@ def test_negative(check: Check):
 @pytest.mark.parametrize(
     "rule_value, rule_data",
     [
-    [(0, 9), np.arange(10)],
-    [("2022-01-01", "2022-02-02"), pd.date_range(start="2022-01-01", end="2022-02-01", freq="D").strftime("%Y-%m-%d").to_list(),],
+        [(0, 9), np.arange(10)],
+        [
+            ("2022-01-01", "2022-02-02"),
+            pd.date_range(start="2022-01-01", end="2022-02-01", freq="D")
+            .strftime("%Y-%m-%d")
+            .to_list(),
+        ],
     ],
     ids=("numeric", "date"),
 )
@@ -40,4 +45,6 @@ def test_coverage(check: Check):
     df = daft.from_pydict({"id": np.arange(20)})
     result = check.validate(df)
     assert result.select(daft.col("status").str.match("PASS")).to_pandas().status.all()
-    assert result.select(daft.col("pass_rate").max() == 0.55).to_pandas().pass_rate.all()
+    assert (
+        result.select(daft.col("pass_rate").max() == 0.55).to_pandas().pass_rate.all()
+    )
