+++ test/unit/daft/test_is_equal_than.py
@@ -32,4 +32,6 @@ def test_coverage(check: Check):
     df = daft.from_pydict({"id": [1, 1, 1, 0]})
     result = check.validate(df)
     assert result.select(daft.col("status").str.match("PASS")).to_pandas().status.all()
-    assert result.select(daft.col("pass_rate").max() == 0.75).to_pandas().pass_rate.all()
+    assert (
+        result.select(daft.col("pass_rate").max() == 0.75).to_pandas().pass_rate.all()
+    )
