+++ test/unit/daft/test_has_workflow.py
@@ -30,5 +30,6 @@ def test_coverage(check: Check):
     )
     result = check.validate(df)
     assert result.select(daft.col("status").str.match("PASS")).to_pandas().status.all()
-    assert result.select(daft.col("pass_rate").max() == 4 / 6).to_pandas().pass_rate.all()
-
+    assert (
+        result.select(daft.col("pass_rate").max() == 4 / 6).to_pandas().pass_rate.all()
+    )

[…]
