+++ .github/workflows/ci.yml
@@ -12,7 +12,7 @@ jobs:
     - name: Setup Python
       uses: actions/setup-python@master
       with:
-        python-version: 3.8
+        python-version: 3.11
     - name: Authenticate google cloud
       uses: 'google-github-actions/auth@v1'
       with:
