### 2024-03-27 9afb5cf
Added contributors image
