+++ pyproject.toml
@@ -25,7 +25,9 @@ dependencies = [
 [project.optional-dependencies]
 dev = [
   "black==24.3.0",
-  "ruff==0.3.3"
+  "ruff==0.3.3",
+  "pytest-postgresql==6.0.0",
+  "psycopg-binary==3.1.18"
 ]
 pyspark = [
   "pyspark>=3.4.0"
@@ -66,6 +68,10 @@ pdf = [
 ]
 daft = [
   "getdaft == 0.2.19"
+]
+db = [
+  "polars>=0.19.6",
+  "connectorx == 0.3.2"
 ]
 
 [tool.ruff.lint]
