+++ test/unit/db/test_is_unique.py
@@ -6,19 +6,20 @@ from cuallee import Check
 def test_positive(check: Check, postgresql, db_conn):
     check.is_unique("id")
     check.table_name = "public.test1"
-    result = check.validate(db_conn).select(pl.col("status")) == "PASS"
-    assert all(result.to_series().to_list())
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
 
 
 def test_negative(check: Check, postgresql, db_conn):
-    check.is_unique("id2")
+    check.is_unique("id3")
     check.table_name = "public.test1"
-    result = check.validate(db_conn).select(pl.col("status")) == "FAIL"
-    assert all(result.to_series().to_list())
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
 
 
 def test_coverage(check: Check, postgresql, db_conn):
-    check.is_unique("id3", 0.5)
+    check.is_unique("id5", 0.5)
     check.table_name = "public.test1"
-    result = check.validate(db_conn).select(pl.col("status")) == "PASS"
-    assert all(result.to_series().to_list())
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+    assert (result.select(pl.col("pass_rate")) == 6/10).to_series().all()
