### 2024-03-28 2bec963
- Added `are_unique`
