### 2024-03-28 96cd185
- Added `is_complete` and `are_complete`
