+++ test/unit/db/test_is_complete.py
@@ -1,0 +1,26 @@
+import pytest
+import polars as pl
+
+from cuallee import Check
+
+
+def test_positive(check: Check, postgresql, db_conn):
+    check.is_complete("id")
+    check.table_name = "public.test1"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+
+
+def test_negative(check: Check, postgresql, db_conn):
+    check.is_complete("id6")
+    check.table_name = "public.test1"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
+
+
+def test_coverage(check: Check, postgresql, db_conn):
+    check.is_complete("id6", 0.5)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+    assert (result.select(pl.col("pass_rate")) == 6/10).to_series().all()
