### 2024-03-28 59becf7
- Added `has_min`, `has_max', `has_sum`, `has_mean`, `has_std`
