+++ test/unit/db/test_has_sum.py
@@ -1,0 +1,18 @@
+import pytest
+import polars as pl
+
+from cuallee import Check
+
+
+def test_positive(check: Check, postgresql, db_conn):
+    check.has_sum("id", 15)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+
+
+def test_negative(check: Check, postgresql, db_conn):
+    check.has_sum("id", 11)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
