+++ pyproject.toml
@@ -27,7 +27,8 @@ dev = [
   "black==24.3.0",
   "ruff==0.3.3",
   "pytest-postgresql==6.0.0",
-  "psycopg-binary==3.1.18"
+  "psycopg-binary==3.1.18",
+#  "pytest-mysql==2.5.0"
 ]
 pyspark = [
   "pyspark>=3.4.0"
