### 2024-03-28 faeec27
- Added `has_correlation`, `has_cardinality`, `has_infogain`
