### 2024-04-02 6a7b70f
- Added `has_pattern`
