### 2024-04-02 41d71fc
- Added `is_equal_than`, `is_less_or_equal_than`, `is_less_than`, `is_greater_or_equal_than`, `is_greater_than`
