+++ test/unit/db/test_is_less_or_equal_than.py
@@ -1,0 +1,26 @@
+import polars as pl
+
+from cuallee import Check
+
+
+def test_positive(check: Check, postgresql, db_conn):
+    check.is_less_or_equal_than("id", 5)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+
+
+def test_negative(check: Check, postgresql, db_conn):
+    check.is_less_or_equal_than("id2", 5)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
+
+
+
+def test_coverage(check: Check, postgresql, db_conn):
+    check.is_less_or_equal_than("id5", 2, 8/10)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+    assert (result.select(pl.col("pass_rate")) == 8/10).to_series().all()
