### 2024-04-02 c02efdc
- Added `has_percentile`
