+++ test/unit/db/test_has_percentile.py
@@ -1,7 +1,7 @@
-import polars as pl
-import numpy as np
-from cuallee import Check
 import pytest
+import polars as pl
+
+from cuallee import Check
 
 
 def test_positive(check: Check, postgresql, db_conn):
@@ -13,7 +13,6 @@ def test_positive(check: Check, postgresql, db_conn):
 
 def test_negative(check: Check, postgresql, db_conn):
     check.has_percentile("id", 3, 0.75)
-    df = pl.DataFrame({"id": np.arange(10)})
     check.table_name = "public.test1"
     result = check.validate(db_conn)
     assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
@@ -23,7 +22,6 @@ def test_values_int(check: Check, postgresql, db_conn):
     check.has_percentile("id2", 8, 0.5)
     check.table_name = "public.test1"
     result = check.validate(db_conn)
-    print(result)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
 
 
