+++ test/unit/db/test_not_contained_in.py
@@ -1,0 +1,26 @@
+import polars as pl
+
+from cuallee import Check
+
+
+def test_positive(check: Check, postgresql, db_conn):
+    check.not_contained_in("id2", [1, 2, 3, 4, 5])
+    check.table_name = "public.test1"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+
+
+def test_negative(check: Check, postgresql, db_conn):
+    check.not_contained_in("id", [1, 2, 3, 4, 5])
+    check.table_name = "public.test1"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
+
+
+def test_coverage(check: Check, postgresql, db_conn):
+    check.not_contained_in( "id5", [0, 1, 2], 0.2)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn)
+    print(result)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+    assert (result.select(pl.col("pass_rate")) == 2/10).to_series().all()
