+++ test/unit/db/test_not_contained_in.py
@@ -21,6 +21,5 @@ def test_coverage(check: Check, postgresql, db_conn):
     check.not_contained_in( "id5", [0, 1, 2], 0.2)
     check.table_name = "public.test1"
     result = check.validate(db_conn)
-    print(result)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
     assert (result.select(pl.col("pass_rate")) == 2/10).to_series().all()
