### 2024-04-03 b3e7b4f
- Added `is_on_schedule`, `is_on_sunday`, `is_on_friday`, `is_on_monday`, `is_on_saturday`, `is_on_thursday`, `is_on_tuesday`, `is_on_wednesday`, `is_on_weekday` and `is_on_weekend`
