+++ test/unit/db/test_is_on_schedule.py
@@ -1,0 +1,28 @@
+import polars as pl
+
+from cuallee import Check
+
+
+def test_positive(check: Check, postgresql, db_conn):
+    check.is_on_schedule("id2", (9, 17))
+    check.table_name = "public.test6"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+
+
+
+def test_negative(check: Check, postgresql, db_conn):
+    check.is_on_schedule("id2", (16, 20))
+    check.table_name = "public.test6"
+    result = check.validate(db_conn)
+    assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
+
+
+def test_coverage(check: Check, postgresql, db_conn):
+    check.is_on_schedule("id2", (12, 17), pct=3/5)
+    check.table_name = "public.test6"
+    result = check.validate(db_conn)
+    print(result)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+    assert (result.select(pl.col("pass_rate")) == 3/5).to_series().all()
+
