### 2024-04-03 712061f
- Added `is_in_millions` and `is_in_billions`
