### 2024-04-03 472e961
- Added `is_inside_interquartile_range`
