### 2024-04-03 7178700
- Added `satisfies`
