### 2024-04-03 203ed55
- Added `is_positive`, `is_negative`
