+++ test/unit/psql/test_is_equal_than.py
@@ -3,24 +3,24 @@ import polars as pl
 from cuallee import Check
 
 
-def test_positive(check: Check, postgresql, db_conn):
+def test_positive(check: Check, postgresql, db_conn_psql):
     check.is_equal_than("id3", 1)
     check.table_name = "public.test1"
-    result = check.validate(db_conn)
+    result = check.validate(db_conn_psql)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
 
 
-def test_negative(check: Check, postgresql, db_conn):
+def test_negative(check: Check, postgresql, db_conn_psql):
     check.is_equal_than("id4", 3)
     check.table_name = "public.test1"
-    result = check.validate(db_conn)
+    result = check.validate(db_conn_psql)
     assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
 
 
-def test_coverage(check: Check, postgresql, db_conn):
+def test_coverage(check: Check, postgresql, db_conn_psql):
     check.is_equal_than("id5", 2, 4/10)
     check.table_name = "public.test1"
-    result = check.validate(db_conn)
+    result = check.validate(db_conn_psql)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
     assert (result.select(pl.col("pass_rate")) == 4/10).to_series().all()
 
