### 2024-04-03 a5fe218
- Added `is_unique` and `are_unique`
