+++ cuallee/mysql_validation.py
@@ -1,0 +1,165 @@
+import operator
+import numpy as np
+import pandas as pd
+import polars as pl
+
+from toolz import first
+from numbers import Number
+from string import Template
+from functools import reduce
+
+from cuallee import Check, Rule, db_connector
+from cuallee.duckdb_validation import Compute as duckdb_compute
+
+import textwrap
+from pygments import highlight
+from pygments.lexers import SqlLexer
+from pygments.formatters.terminal256 import TerminalTrueColorFormatter
+
+
+class Compute(duckdb_compute):
+
+    def __init__(self, table_name: str = None):
+        super().__init__(table_name)
+
+    def are_unique(self, rule: Rule) -> str:
+        """Validate absence of duplicate in group of columns"""
+        return "( "+ " + ".join( f"COUNT(DISTINCT({column}))" for column in rule.column) + f" ) / {float(len(rule.column))} "
+
+    def has_std(self, rule: Rule) -> str:
+        """Validate standard deviation on column"""
+        #BUG: This could fail due to floating point precision
+        #IDEA: Use f"CAST(STDDEV_SAMP({rule.column}) AS FLOAT) - CAST({rule.value} AS FLOAT) < {percision_error}"
+        return f"CAST(STDDEV_SAMP({rule.column}) AS FLOAT) = CAST({rule.value} AS FLOAT)"
+
+    def has_entropy(self, rule: Rule) -> str:
+        """Computes entropy of 0-1 vector."""
+        raise NotImplementedError
+
+    def has_max_by(self, rule: Rule) -> str:
+        """Adjacent column maximum value verifiation on threshold"""
+        """
+        ```sql
+        SELECT id
+        FROM public.test1
+        WHERE id2 = (SELECT MAX(id2) FROM public.test1);
+        ```
+        """
+        raise NotImplementedError
+
+    def has_min_by(self, rule: Rule) -> str:
+        """Adjacent column minimum value verifiation on threshold"""
+        """
+        ```sql
+        SELECT id
+        FROM public.test1
+        WHERE id2 = (SELECT MIN(id2) FROM public.test1);
+        ```
+        """
+        raise NotImplementedError
+
+    def has_percentile(self, rule: Rule) -> str:
+        """Percentile range verification for column"""
+        return f"PERCENTILE_CONT({rule.settings['percentile']}) WITHIN GROUP (ORDER BY {rule.column})  = {rule.value}"
+
+    def is_inside_interquartile_range(self, rule: Rule) -> str:
+        """Validates a number resides inside the Q3 - Q1 range of values"""
+        return f"""
+                SUM(CASE WHEN {rule.column}
+                BETWEEN
+                (SELECT PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY {rule.column}) AS first_quartile FROM {self.table_name})
+                AND
+                (SELECT PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY {rule.column}) AS third_quartile FROM {self.table_name})
+                THEN 1 ELSE 0 END)
+                """
+
+    def has_workflow(self, rule: Rule) -> str:
+        """Validates events in a group clause with order, followed a specific sequence. Similar to adjacency matrix validation"""
+        raise NotImplementedError
+
+def validate_data_types(check: Check, dataframe):
+    return True
+
+def compute(check: Check):
+    return True
+
+
+def summary(check: Check, connection: db_connector) -> list:
+    unified_columns = ",\n\t".join(
+        [
+            # This is the same as compute.`rule.method`(rule)
+            operator.methodcaller(rule.method, rule)(Compute(check.table_name))
+            + f' AS "{rule.key}"'
[…]
