### 2024-04-04 0861b5e
- Added `has_cardinality`
