+++ test/unit/psql/test_has_cardinality.py
@@ -6,11 +6,11 @@ from cuallee import Check
 def test_positive(check: Check, postgresql, db_conn_psql):
     check.has_cardinality("id", 5)
     check.table_name = "public.test1"
-    result = check.validate(db_conn_psql).select(pl.col("status")) == "PASS"
-    assert all(result.to_series().to_list())
+    result = check.validate(db_conn_psql)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
 
 def test_negative(check: Check, postgresql, db_conn_psql):
     check.has_cardinality("id", 4)
     check.table_name = "public.test1"
-    result = check.validate(db_conn_psql).select(pl.col("status")) == "FAIL"
-    assert all(result.to_series().to_list())
+    result = check.validate(db_conn_psql)
+    assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
