+++ test/unit/mysql/test_has_cardinality.py
@@ -1,0 +1,37 @@
+import polars as pl
+
+from cuallee import Check
+
+# CHECK: results if they are correct
+
+def test_positive(check: Check, db_conn_mysql):
+    check.has_cardinality("id", 5)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn_mysql)
+    print(result)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+
+
+def test_negative(check: Check, db_conn_mysql):
+    check.has_cardinality("id", 3)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn_mysql)
+    print(result)
+    assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
+
+
+def test_edge_case(check: Check, db_conn_mysql):
+    """A column that have cardinality of 3"""
+    check.has_cardinality("id5", 3)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn_mysql)
+    print(result)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+
+def test_edge_case_null(check: Check, db_conn_mysql):
+    """A column will nulls"""
+    check.has_cardinality("id6", 3)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn_mysql)
+    print(result)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
