+++ test/unit/mysql/first_test.py
@@ -4,3 +4,4 @@ def test_mysql_docker(db_conn_mysql):
     assert db_conn_mysql(query="SELECT count(*) FROM public.test1").item(0,0) == 5
     assert db_conn_mysql(query="SELECT count(*) FROM public.test3").item(0,0) == 6
     assert db_conn_mysql(query="SELECT count(*) FROM public.test7").item(0,0) == 3
+    assert db_conn_mysql(query="SELECT COUNT(DISTINCT(id)) FROM public.test1").item(0,0) == 5
