### 2024-04-04 6ea4c35
- Added `has_infogain`
