+++ test/unit/mysql/test_has_correlation.py
@@ -3,12 +3,13 @@ import polars as pl
 
 from cuallee import Check
 
+# [ ]: has_correlation
+
 @pytest.mark.skip(reason="Not implemented yet!")
 def test_positive(check: Check, db_conn_mysql):
     check.has_correlation("id", "id2", 1.0)
     check.table_name = "public.test2"
     result = check.validate(db_conn_mysql)
-    print(result)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
 
 @pytest.mark.skip(reason="Not implemented yet!")
