### 2024-04-04 c0f24ac
- Added `has_max`, `has_mean`, `has_min`, `has_sum`, , `has_std`
