+++ test/unit/mysql/test_has_entropy.py
@@ -6,18 +6,18 @@ from cuallee import Check
 # [ ]: has_entropy
 
 @pytest.mark.skip(reason="Not implemented yet!")
-def test_positive(check: Check, postgresql, db_conn_psql):
+def test_positive(check: Check, db_conn_mysql):
     check.has_entropy("id", 1.0)
     check.table_name = "public.test3"
-    result = check.validate(db_conn_psql)
+    result = check.validate(db_conn_mysql)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
 
 @pytest.mark.skip(reason="Not implemented yet!")
-def test_negative(check: Check, postgresql, db_conn_psql):
+def test_negative(check: Check, db_conn_mysql):
     check.has_entropy("id", 1.0)
     df = pl.DataFrame({"id": [10, 10, 10, 10, 50]})
     check.table_name = "public.test1"
-    result = check.validate(db_conn_psql)
+    result = check.validate(db_conn_mysql)
     assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
 
 
@@ -25,7 +25,7 @@ def test_negative(check: Check, postgresql, db_conn_psql):
 @pytest.mark.parametrize(
     "values", [[1], [1, 1, 1, 1, 1]], ids=("observation", "classes")
 )
-def test_parameters(check: Check, values, postgresql, db_conn_psql):
+def test_parameters(check: Check, values, db_conn_mysql):
     check.has_entropy("id", 0.0)
     df = pl.DataFrame({"id": values})
     result = check.validate(df).select(pl.col("status")) == "PASS"
@@ -33,6 +33,6 @@ def test_parameters(check: Check, values, postgresql, db_conn_psql):
 
 
 @pytest.mark.skip(reason="Not implemented yet!")
-def test_coverage(check: Check, postgresql, db_conn_psql):
+def test_coverage(check: Check, db_conn_mysql):
     with pytest.raises(TypeError):
         check.has_entropy("id", 1.0, pct=0.5)
