+++ test/unit/mysql/test_has_sum.py
@@ -1,0 +1,17 @@
+import polars as pl
+
+from cuallee import Check
+
+
+def test_positive(check: Check, db_conn_mysql):
+    check.has_sum("id", 15)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn_mysql)
+    assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
+
+
+def test_negative(check: Check, db_conn_mysql):
+    check.has_sum("id", 11)
+    check.table_name = "public.test1"
+    result = check.validate(db_conn_mysql)
+    assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
