### 2024-04-04 ed18f3b
- Added `is_complete`
