### 2024-04-04 7fb1623
- Added `is_equal_than`, `is_greater_or_equal_than`, `is_greater_than`, `is_in_billions`, `is_in_millions`, `is_less_or_equal_than`, `is_less_than`
