### 2024-04-04 d452273
- Added `is_positive` and `is_negative`
