+++ cuallee/mysql_validation.py
@@ -109,6 +109,10 @@ class Compute(duckdb_compute):
     def is_contained_in(self, rule: Rule) -> str:
         return f"SUM({rule.column} IN {rule.value})"
 
+    def not_contained_in(self, rule: Rule) -> str:
+        """Validation of column value not in a set of given values"""
+        return f"SUM({rule.column} NOT IN {rule.value})"
+
     def is_inside_interquartile_range(self, rule: Rule) -> str:
         """Validates a number resides inside the Q3 - Q1 range of values"""
         raise NotImplementedError
