### 2024-04-04 d4abe13
- Added `satisfies`
