### 2024-04-04 f2c7bfb
- Added `is_on_weekend`, `is_on_friday`, `is_on_monday`, `is_on_saturday`, `is_on_sunday`, `is_on_thursday`, `is_on_tuesday`, `is_on_wednesday`, `is_on_weekday`
