### 2024-04-04 067a866
- Added `has_pattern`
