+++ test/unit/mysql/test_has_max_by.py
@@ -29,7 +29,7 @@ def test_negative(check: Check, postgresql, db_conn_psql):
 )
 def test_values(check: Check, answer, columns, postgresql, db_conn_psql):
     check.has_max_by("id", "id2", answer)
-    df = pl.DataFrame({"id": columns, "id2": [300, 500]})
+    df = pl.DataFrame({"id": columns, "id2": [300, 500]})  # noqa: F841
     check.table_name = "public.test1"
     result = check.validate(db_conn_psql)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
