+++ test/unit/mysql/test_is_on_monday.py
@@ -7,7 +7,6 @@ def test_positive(check: Check, db_conn_mysql):
     check.is_on_monday("id2", pct= 0.20)
     check.table_name = "public.test5"
     result = check.validate(db_conn_mysql)
-    print(result)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
 
 
@@ -22,6 +21,5 @@ def test_coverage(check: Check, db_conn_mysql):
     check.is_on_monday("id2", pct= 2/9)
     check.table_name = "public.test5"
     result = check.validate(db_conn_mysql)
-    print(result)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
     assert (result.select(pl.col("pass_rate")) == 2/9).to_series().all()
