+++ test/unit/mysql/test_has_cardinality.py
@@ -7,7 +7,6 @@ def test_positive(check: Check, db_conn_mysql):
     check.has_cardinality("id", 5)
     check.table_name = "public.test1"
     result = check.validate(db_conn_mysql)
-    print(result)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
 
 
@@ -15,7 +14,6 @@ def test_negative(check: Check, db_conn_mysql):
     check.has_cardinality("id", 3)
     check.table_name = "public.test1"
     result = check.validate(db_conn_mysql)
-    print(result)
     assert (result.select(pl.col("status")) == "FAIL" ).to_series().all()
 
 
@@ -24,7 +22,6 @@ def test_edge_case(check: Check, db_conn_mysql):
     check.has_cardinality("id5", 3)
     check.table_name = "public.test1"
     result = check.validate(db_conn_mysql)
-    print(result)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
 
 def test_edge_case_null(check: Check, db_conn_mysql):
@@ -32,5 +29,4 @@ def test_edge_case_null(check: Check, db_conn_mysql):
     check.has_cardinality("id6", 3)
     check.table_name = "public.test1"
     result = check.validate(db_conn_mysql)
-    print(result)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
