+++ test/unit/psql/test_is_on_friday.py
@@ -7,7 +7,6 @@ def test_positive(check: Check, postgresql, db_conn_psql):
     check.is_on_friday("id2", pct= 0.10)
     check.table_name = "public.test5"
     result = check.validate(db_conn_psql)
-    print(result)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
 
 
@@ -22,6 +21,5 @@ def test_coverage(check: Check, postgresql, db_conn_psql):
     check.is_on_friday("id2", pct= 1/9)
     check.table_name = "public.test5"
     result = check.validate(db_conn_psql)
-    print(result)
     assert (result.select(pl.col("status")) == "PASS" ).to_series().all()
     assert (result.select(pl.col("pass_rate")) == 1/9).to_series().all()
