+++ test/unit/daft/test_has_workflow.py
@@ -1,6 +1,4 @@
 import daft
-import pytest
-import numpy as np
 
 from cuallee import Check
 
