+++ cuallee/psql_validation.py
@@ -152,9 +152,9 @@ def summary(check: Check, connection: db_connector) -> list:
             "timestamp": check.date.strftime("%Y-%m-%d %H:%M:%S"),
             "check": check.name,
             "level": check.level.name,
-            "column": rule.column,
+            "column": str(rule.column),
             "rule": rule.method,
-            "value": rule.value,
+            "value": str(rule.value),
             "rows": rows,
             "violations": _calculate_violations(first(unified_results[hash_key[:-1]]), rows),
             "pass_rate": _calculate_pass_rate(first(unified_results[hash_key[:-1]]), rows),
