+++ README.md
@@ -14,7 +14,7 @@
 
 Meaning `good` in Aztec (Nahuatl), _pronounced: QUAL-E_
 
-This library provides an intuitive `API` to describe `checks` initially just for `PySpark` dataframes `v3.3.0`. And extended to `pandas`, `snowpark`, `duckdb`, `daft` and more.
+This library provides an intuitive `API` to describe `checks` initially just for `PySpark` dataframes `v3.3.0`. And extended to `Pandas`, `Snowpark`, `DuckDB`, `BigQuery` ,`Daft`, `MySQL`, `PostgreSQL` and more.
 It is a replacement written in pure `python` of the `pydeequ` framework.
 
 I gave up in _deequ_ as after extensive use, the API is not user-friendly, the Python Callback servers produce additional costs in our compute clusters, and the lack of support to the newest version of PySpark.
@@ -38,6 +38,8 @@ Provider | API | Versions
 ![duckdb](logos/duckdb.png?raw=true "DuckDB API")|`duckdb` | `0.9.2`,~~`0.8.0`~~, ~~`0.7.1`~~
 ![polars](logos/polars.svg?raw=true "Polars API")|`polars`| `0.19.6`
 ![daft](logos/daft.png?raw=true "Daft API")|`daft`| `0.2.19`
+![postgresql](logos/postgres.png?raw=true "PostgreSQL")|`postgresql`|
+![mysql](logos/mysql.png?raw=true "MySQL")|`mysql`|
 
  <sub>Logos are trademarks of their own brands.</sub>
 
@@ -313,6 +315,104 @@ check.validate(conn)
 1   2  2022-10-31 23:15:06  test  WARNING  tpep_pickup_datetime  is_complete   N/A  19817583         0.0        1.0             1.0   PASS
 ```
 
+## Relational Databases
+It is possible to use `cuallee` to validate tables in relational databases.
+
+### PostgreSQL
+
+All the checks are not available for `PostgreSQL` at the moment.
+List of unavailable checks:
+- `is_daily`
+- `has_entrop`
+- `has_max_by`
+- `has_min_by`
+- `has_workflow`
+
+<br>
+
+```python
+from cuallee import db_connector, Check, CheckLevel
+
+# Postgres URI Connection
+uri = "postgresql://<User>:<Password>@<Host>/<Database>"
+
+# Using db_connector to create connector to the database
+db_conn_psql = db_connector(uri)
+
+# Create Check
+check = Check(CheckLevel.WARNING, "PostgresCheck")
+
+# Table Name with Schema
+check.table_name = "schema.table"
+
+(  check.are_complete(("id", "id2"))
+        .is_positive("id", 0.5)
+        .is_less_than("id", 50)
+        .is_between("id2", (0, 100)))
+
+check.validate(db_conn_psql)
+
+┌─────┬─────────────────────┬───────────────┬─────────┬───────────────┬─────────────────┬──────────┬──────┬────────────┬───────────┬────────────────┬────────┐
+│ id  ┆ timestamp           ┆ check         ┆ level   ┆ column        ┆ rule            ┆ value    ┆ rows ┆ violations ┆ pass_rate ┆ pass_threshold ┆ status │
+│ --- ┆ ---                 ┆ ---           ┆ ---     ┆ ---           ┆ ---             ┆ ---      ┆ ---  ┆ ---        ┆ ---       ┆ ---            ┆ ---    │
[…]
