+++ pyproject.toml
@@ -28,7 +28,7 @@ dev = [
   "ruff==0.3.3",
   "pytest-postgresql==6.0.0",
   "psycopg-binary==3.1.18",
-#  "pytest-mysql==2.5.0"
+  "pytest-mysql==2.5.0"
 ]
 pyspark = [
   "pyspark>=3.4.0"
