+++ cuallee/mysql_validation.py
@@ -31,9 +31,8 @@ class Compute(duckdb_compute):
     def has_std(self, rule: Rule) -> str:
         """Validate standard deviation on column.
         Issues: #Note: This could fail due to floating point precision
-        Suggestions: #IDEA: Use f"CAST(STDDEV_SAMP({rule.column}) AS FLOAT) - CAST({rule.value} AS FLOAT) < {percision_error}"
+        Suggestions: #IDEA: Use f"IF( CAST(STDDEV_SAMP({rule.column}) AS FLOAT) - CAST({rule.value} AS FLOAT) < {percision_error}, 'True', 'False')"
         """
-
         return f"IF( CAST(STDDEV_SAMP({rule.column}) AS FLOAT) = CAST({rule.value} AS FLOAT), 'True', 'False')"
 
     def has_mean(self, rule: Rule) -> str:
