### 2024-04-05 b0cf1dc
🧹 Removed unnecessary inheritance
