### 2024-04-05 2fbd5d2
➕ Added `has_max_by` and `has_min_by` to mysql
