### 2024-04-05 6c6790f
➕ Added `has_min_by` and `has_max_by` for psql
