### 2024-04-08 be114e8
Bump getdaft from 0.2.19 to 0.2.20
Bumps [getdaft](https://github.com/Eventual-Inc/Daft) from 0.2.19 to 0.2.20.
- [Release notes](https://github.com/Eventual-Inc/Daft/releases)
- [Commits](https://github.com/Eventual-Inc/Daft/compare/v0.2.19...v0.2.20)

---
updated-dependencies:
- dependency-name: getdaft
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
