+++ pyproject.toml
@@ -25,7 +25,7 @@ dependencies = [
 [project.optional-dependencies]
 dev = [
   "black==24.3.0",
-  "ruff==0.3.3"
+  "ruff==0.3.5"
 ]
 pyspark = [
   "pyspark>=3.4.0"
