### 2024-04-10 68ee0b9
remodeled the cuallee to compute all rules in this check for specific data frame
