### 2024-04-22 ea93035
Bump duckdb from 0.10.0 to 0.10.2
Bumps [duckdb](https://github.com/duckdb/duckdb) from 0.10.0 to 0.10.2.
- [Release notes](https://github.com/duckdb/duckdb/releases)
- [Changelog](https://github.com/duckdb/duckdb/blob/main/tools/release-pip.py)
- [Commits](https://github.com/duckdb/duckdb/compare/v0.10.0...v0.10.2)

---
updated-dependencies:
- dependency-name: duckdb
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
