### 2024-04-24 #6459 @jbytecode
👋👋👋 Dear @FlorianK13, @SarthakJariwala 👋👋👋

**[cuallee: A python package for data quality checks across multiple DataFrame APIs](https://github.com/openjournals/joss-reviews/issues/6459)**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/6459).

Thank you in advance!
