### 2024-04-29 canimus/cuallee#212 @devarops
[JOSS REVIEW] Automatic testing is failing
Hi @canimus,

The test `snowpark_dataframe/test_are_complete.py::test_positive` ends with an error.

``` python
ERROR test/unit/snowpark_dataframe/test_are_complete.py::test_positive - ValueError: snowpark did not yield a value
```

Please pass the tests in [this pipeline](https://github.com/canimus/cuallee/actions/workflows/ci.yml).

- https://github.com/openjournals/joss-reviews/issues/6684
