### 2024-04-30 canimus/cuallee#212 @canimus
Hi @devarops in order to test the integration scenarios, is required to provide a SNOWFLAKE account through a series of environment variables:
- `SF_ACCOUNT`
- `SF_USER`
- `SF_PASSWORD`
- `SF_ROLE`
- `SF_WAREHOUSE`
- `SF_DATABASE`
- `SF_SCHEMA`

Same will happen with BigQuery, a GOOGLE_CREDENTIALS file with the `json` key is required. Perhaps we should rather move this tests, into an integration folder, and leave the `unit` one to those can be executed in full isolation.
As the submission made initial reference for the `pyspark` solution, and in fact the rest of the tests are simply organized as we did the ones for Spark, then running `pytest test/unit/pyskark_dataframe`
