### 2024-04-30 canimus/cuallee#213 @devarops
[JOSS REVIEW] Community guidelines
Hi @canimus,

I can see a link to the [contributing guidelines](https://github.com/canimus/cuallee/blob/main/CONTRIBUTING.md) on a sidebar titled **Helpful resources** when creating a new issue.

However, I need help finding a link to these guidelines within the documentation.
Can you please point me to the section in the README with the guidelines for third parties wishing to:

1. Contribute to the software.
1. Report issues or problems with the software.
1. Seek support

Thank you!

- https://github.com/openjournals/joss-reviews/issues/6684
