### 2024-04-30 canimus/cuallee#213 @canimus
Hi @devarops thanks for the due diligence, greatly appreciated.
- CONTRIBUTING.md under the root folder provides the minimal steps to participate in this open source project. As you can see from the Contributors section of the README.md we gladly have welcomed commits from the community, and receive overwhelming reception and usage as well, with stars and mentions in HackerNews.
- Under the `.github/ISSUE_TEMPLATE/bug_report.md` file, there is a pre-defined template with guidelines to submit an issue. Examples of issues opened/closed are already available in the repository. We have a brief, yet effective approach to define the defects of the software. Gladly the `98%` code coverage also helps preserve the quality of it over time.
- As an open source software, we rely 100% on the community to provide support, at this point in our semantic versioning strategy, we haven't push the envelop to `1.0.0` so that we can at least reach a stable, robust and durable platform. That being said, as stated in the CONTRIBUTING.md you can reach out to `canimus@gmail.com` with questions that I will do my best to answer promptly, and alternatively, open a discussion in the repo, which is open to everyone, to seek for prompt support or Q/A feedback loops.

I hope the references provided suffice to cover the inquiry above.
