### 2024-04-30 1fd18c6
Bump duckdb and fixed teste cases
