+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
 name = cuallee
-version = 0.10.0
+version = 0.10.1
 [options]
 packages = find:
