### 2024-04-30 dccf740
Bump duckdb and fixed teste cases (#215)
