### 2024-05-02 canimus/cuallee#216 @FlorianK13
[JOSS Review] Extend Docstring of central class
The class `Check` seems like the central entry point to the package, however it lacks a documentation. Could you add docstrings to it to explain the purpose of the function and of all parameters?

In case you have not worked with docstrings yet, they are super helpful in modern IDEs to get info about the function when hovering over it. There are some docstring [format conventions](https://numpydoc.readthedocs.io/en/latest/format.html) and also tools like [autodocstring](https://github.com/NilsJPWerner/autoDocstring) that help to create docstrings.

This issue is part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6684)
