### 2024-05-02 canimus/cuallee#217 @FlorianK13
[JOSS Review] Documentation page
There is a documentation page at https://canimus.github.io/cuallee/ hosting all the files in the `docs` folder.
However it seems to be not quite finished. Is it part of the official documentation? Or is the `Readme.md` the official documentation?

This issue is part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6684)
