### 2024-05-03 canimus/cuallee#217 @canimus
Hi @FlorianK13 thanks for the review process. We started with the idea of building an external site, through Github Pages to host the documentation. However, we decided that investing time in more testing could render better results. As you highlighted, the README.md contains a comprehensive view of all checks available today, and the unit tests, are our best demonstration of the functionality. Nevertheless and without excuses, we should put the references we have in the docs up to date, so that it represent the state of the project, and indeed, can help guide the community.
