### 2024-05-03 canimus/cuallee#216 @canimus
Hi @FlorianK13 thanks for the references. We will act on this, and document more specifically the `Check` class in `cuallee`. Regards H.
