### 2024-05-07 canimus/cuallee#217 @FlorianK13
Hi @canimus, I have to say that I really struggle with the documentation in the Readme.

First a more general remark: Your software seems to have lots of features, i.e. lots of test cases in combination with several different data frameworks. This is (of course) a great thing. Putting the documentation of such an extended software just in one Readme file seems complicated. At least you would need a very clear structure and some sort of navigation like a table of content at the beginning for easier navigation.

Regarding the documentation in the Readme: You start by defining [supported frameworks](https://github.com/canimus/cuallee?tab=readme-ov-file#support) and the [installation](https://github.com/canimus/cuallee?tab=readme-ov-file#install). Then you use one sentence to introduce [Checks](https://github.com/canimus/cuallee?tab=readme-ov-file#checks). I assume that checks are the central entrypoint and that all tests follow this schema
`check.test1("column").test2("column").validate(df).show()` - but this is not documented, so maybe I'm wrong?

After that, some tests are mentioned and some example code is given. When comparing these descriptions with the [Catalogue](https://github.com/canimus/cuallee?tab=readme-ov-file#catalogue) it seems like not all tests are described. Just as an example, `has_entropy` takes three arguments `column`, `value`, and `tolerance`. When using this test, I would need to guess what they represent. A solution could also be docstring documentation here. And [mkdocstrings](https://github.com/mkdocstrings/mkdocstrings) could turn your code documentation in a documentation page - but this is just how I would go here, if you feel more comfortable with other solutions that is also fine.

You can also check the review criteria for documentation pages at the JOSS [website](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation)
