### 2024-05-07 canimus/cuallee#217 @canimus
Hi @FlorianK13 , we opted to put more effort in bringing the docs site alive, with accurate references:
https://canimus.github.io/cuallee/
That hopefully will help to meet the criteria, and provide the structure you very well are suggesting to adhere to, so that navigating the package functionality does not represent a challenge for the end user.
Thanks for your time and patience, on the resolution, hopefully by next week, we will have a docs version up to date, with all the `Checks` up to date. Warm regards, Herminio
