### 2024-05-11 f94a89a
Working documentations with Check and Rule values
