+++ docs/dependencies.md
@@ -1,1 +1,7 @@
 # Dependencies
+
+It relies has only `2` dependencies that are automatically installed with the package:
+
+- `toolz` for functional programming adoption
+- `requests` for downloading ISO standard required by `check.iso` interface.
+- `pandas` as the default validation result data abstraction
