+++ docs/module/rule.md
@@ -1,0 +1,5 @@
+::: cuallee.Rule
+    handler: python
+    rendering:
+        show_bases: False
+        show_inheritance_diagram: True
