+++ .github/workflows/ci.yml
@@ -49,5 +49,5 @@ jobs:
       - uses: actions/setup-python@v4
         with:
           python-version: 3.x
-      - run: pip install mkdocs-material 
+      - run: pip install mkdocs-material mkdocstrings[python]
       - run: mkdocs gh-deploy --force
