+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
 name = cuallee
-version = 0.10.1
+version = 0.10.2
 [options]
 packages = find:

[…]
