### 2024-05-11 56aa49b
Feature docs v2 (#221)
* Added new version of the flow

* Working documentations with Check and Rule values
