+++ cuallee/__init__.py
@@ -12,7 +12,7 @@ from toolz import compose, valfilter  # type: ignore
 from toolz.curried import map as map_curried
 
 logger = logging.getLogger("cuallee")
-__version__ = "0.10.1"
+__version__ = "0.10.2"
 # Verify Libraries Available
 # ==========================
 try:
@@ -57,11 +57,13 @@ except (ModuleNotFoundError, ImportError):
 
 
 class CheckLevel(enum.Enum):
+    """Level of verifications in cuallee"""
     WARNING = 0
     ERROR = 1
 
 
 class CheckDataType(enum.Enum):
+    """Accepted data types in checks"""
     AGNOSTIC = 0
     NUMERIC = 1
     STRING = 2
@@ -71,6 +73,7 @@ class CheckDataType(enum.Enum):
 
 
 class CheckStatus(enum.Enum):
+    """Validation result criteria"""
     PASS = "PASS"
     FAIL = "FAIL"
     NO_RUN = "NO_RUN"
@@ -78,6 +81,7 @@ class CheckStatus(enum.Enum):
 
 @dataclass
 class Rule:
+    """Predicate definition holder"""
     method: str
     column: Union[str, List[str], Tuple[str, str]]
     value: Optional[Any]
@@ -91,10 +95,12 @@ class Rule:
 
     @property
     def settings(self) -> dict:
+        """holds the additional settings for the predicate execution"""
         return dict(self.options)
 
     @property
     def key(self):
+        """blake2s hash of the rule, made of method, column, value, options and coverage"""
         return (
             hashlib.blake2s(
                 bytes(
@@ -182,6 +188,7 @@ class Rule:
 
 
 class ComputeEngine(Protocol):
+    """An interface for validatiosn to adhere to"""
     def compute(self, rules: Dict[str, Rule]) -> bool:
         """Returns compute instructions for each rule"""
 
@@ -202,7 +209,17 @@ class Check:
         table_name: str = None,
         session: Any = None,
     ):
-        """A container of data quality rules."""
+        """
+        A container of data quality rules.
+
+        Args:
+            level (CheckLevel): [0-1] value to describe if its a WARNING or ERROR check
+            name (str): Normally the name of the dataset being verified, or a name for this check
+            execution_date (date): An automatically generated timestamp of the check in UTC
+            table_name (str): When using databases matches the table name of the source 
+            session (Session): When operating in Session enabled environments like Databricks or Snowflake
+        
+        """
         self._rule: Dict[str, Rule] = {}
         # TODO: Should be a compute engine protocol
         self.compute_engine: ModuleType
@@ -233,7 +250,7 @@ class Check:
 
     @property
     def sum(self):
-        """Collect compute, unique and union type of rules"""
+        """Total number of rules in Check"""
         return len(self._rule.keys())
 
     @property
@@ -252,16 +269,32 @@ class Check:
         return len(self.rules) == 0
 
     def _remove_rule_generic(self, key: str):
-        """Remove a key from rules and compute dictionaries"""
+        """
+        Remove a key from rules and compute dictionaries
+        
+        Args:
+            key (str): the blake2s key of the rule
+        """
         if key in self._rule:
             self._rule.pop(key)
 
     def add_rule(self, method: str, *arg):
-        """Add a new rule to the Check class."""
+        """
+        Add a new rule to the Check class.
+        
+        Args:
+            method (str): Check name
+            args (list): Parameters of the check
+        """
         return operator.methodcaller(method, *arg)(self)
 
     def delete_rule_by_key(self, keys: Union[str, List[str]]):
-        """Delete rules from self._rule and self._compute dictionnary based on keys."""
+        """
+        Delete rules from check based on keys.
+        
+        Args:
[…]
