+++ docs/new-check.md
@@ -1,1 +1,0 @@
-# New Check
