+++ docs/module/check.md
@@ -1,0 +1,10 @@
+
+::: cuallee.Check
+    handler: python
+    rendering:
+        show_bases: False
+        heading_level: 3
+
+
+
+        
