### 2024-05-11 c380df8
Bumped version and added README reference to documentation
