### 2024-05-11 canimus/cuallee#212 @canimus
@devarops there is a free tier for a Snowflake account that expires every month. Credentials in the Github actions have to be renewed every month, based on this limitation. In principle, unit tests should mock the interfaces but we found it more representative, to actually run the tests again the real systems, specially if there is not an additional costs to do so.
So the fix on the pipeline, was literally just changing the credential references to a new "active" trial account in Snowflake.
