+++ pyproject.toml
@@ -25,7 +25,7 @@ dependencies = [
 [project.optional-dependencies]
 dev = [
   "black==24.4.2",
-  "ruff==0.4.3"
+  "ruff==0.4.4"
 ]
 pyspark = [
   "pyspark>=3.4.0"
