### 2024-05-14 canimus/cuallee#228 @FlorianK13
[JOSS Review] `Paper` Report on state of the field
JOSS wants the software papers to contain a comparison to the state of the field, i.e.

```State of the field: Do the authors describe how this software compares to other commonly-used packages?```

Hence, could you add a section in your paper where you shortly compare `cuallee` to other data testing frameworks, like soda, great_expectations, or others?

This issue is part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6684)
