### 2024-05-16 canimus/cuallee#229 @FlorianK13
[JOSS Review] duckdb v0.9.2 does not work
**Describe the bug**
With duckdb v0.9.2 I get the error:
`Exception: Cuallee is not ready for this data structure. You can log a Feature Request in Github.`

**To Reproduce**
Steps to reproduce the behavior:
1. Install duckdb v0.9.2
2. Read the taxi parquet file to duckdb and pandas
3. Define some check
4. Make sure that the check runs sucessfully for the pandas df
5. Try to run the check with duckdb, get the error

 - OS: Windows
 - python: 3.11.9
 - cuallee: 0.10.2

This issue is part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6684)
