### 2024-05-16 91e2fb9
Removed compatibility with older versions of DuckDB
