### 2024-05-16 canimus/cuallee#229 @canimus
Sharing a working example:
https://colab.research.google.com/drive/1-5Bg2L8uF7eX0Jl1qIXznSMtzuZTHrDQ?usp=sharing
