+++ test/unit/daft/test_has_pattern.py
@@ -47,6 +47,7 @@ def test_coverage(check: Check):
     df = daft.from_pydict({"id": ["Herminio", "Hola", "Villain", "Heroe"]})
     result = check.validate(df)
     assert result.select(daft.col("status").str.match("PASS")).to_pandas().status.all()
+    col_pass_rate = daft.col("pass_rate")
     assert (
-        result.select(daft.col("pass_rate").max() == 0.75).to_pandas().pass_rate.all()
+        result.agg(col_pass_rate.max()).select(col_pass_rate == 0.75).to_pandas().pass_rate.all()
     )
