+++ test/unit/daft/test_is_equal_than.py
@@ -32,6 +32,8 @@ def test_coverage(check: Check):
     df = daft.from_pydict({"id": [1, 1, 1, 0]})
     result = check.validate(df)
     assert result.select(daft.col("status").str.match("PASS")).to_pandas().status.all()
+    col_pass_rate = daft.col("pass_rate")
     assert (
-        result.select(daft.col("pass_rate").max() == 0.75).to_pandas().pass_rate.all()
+        result.agg(col_pass_rate.max()).select(col_pass_rate == 0.75).to_pandas().pass_rate.all()
     )
+
