+++ cuallee/daft_validation.py
@@ -19,16 +19,17 @@ from cuallee import Check, Rule
 class Compute:
 
     def is_complete(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
-        perdicate = daft.col(rule.column).not_null().cast(daft.DataType.int64()).sum()
-        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+        col_name = daft.col(rule.column)
+        perdicate =col_name.not_null().cast(daft.DataType.int64())
+        return dataframe.select(perdicate).sum(col_name).to_pandas().iloc[0, 0]
 
     def are_complete(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
         col_names = rule.column
         perdicate = [
-            daft.col(col_name).not_null().cast(daft.DataType.int64()).sum()
+            daft.col(col_name).not_null().cast(daft.DataType.int64())
             for col_name in col_names
         ]
-        return dataframe.select(*perdicate).to_pandas().astype(int).sum().sum() / len(
+        return dataframe.select(*perdicate).sum(col_names).to_pandas().astype(int).sum().sum() / len(
             col_names
         )
 
@@ -52,54 +53,59 @@ class Compute:
     def is_greater_than(
         self, rule: Rule, dataframe: daft.DataFrame
     ) -> Union[bool, int]:
+        col_name = daft.col(rule.column)
         perdicate = (
-            (daft.col(rule.column) > rule.value).cast(daft.DataType.int64()).sum()
+            (col_name > rule.value).cast(daft.DataType.int64())
         )
-        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+        return dataframe.select(perdicate).sum(col_name).to_pandas().iloc[0, 0]
 
     def is_greater_or_equal_than(
         self, rule: Rule, dataframe: daft.DataFrame
     ) -> Union[bool, int]:
+        col_name = daft.col(rule.column)
         perdicate = (
-            (daft.col(rule.column) >= rule.value).cast(daft.DataType.int64()).sum()
+            (col_name >= rule.value).cast(daft.DataType.int64())
         )
-        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+        return dataframe.select(perdicate).sum(col_name).to_pandas().iloc[0, 0]
 
     def is_less_than(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        col_name = daft.col(rule.column)
         perdicate = (
-            (daft.col(rule.column) < rule.value).cast(daft.DataType.int64()).sum()
+            (col_name < rule.value).cast(daft.DataType.int64())
         )
-        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+        return dataframe.select(perdicate).sum(col_name).to_pandas().iloc[0, 0]
 
     def is_less_or_equal_than(
         self, rule: Rule, dataframe: daft.DataFrame
     ) -> Union[bool, int]:
+        col_name = daft.col(rule.column)
         perdicate = (
-            (daft.col(rule.column) <= rule.value).cast(daft.DataType.int64()).sum()
+            (col_name <= rule.value).cast(daft.DataType.int64())
         )
-        return dataframe.select(perdicate).to_pandas().iloc[0, 0]
+        return dataframe.select(perdicate).sum(col_name).to_pandas().iloc[0, 0]
 
     def is_equal_than(self, rule: Rule, dataframe: daft.DataFrame) -> Union[bool, int]:
+        col_name = daft.col(rule.column)
         perdicate = (
-            (daft.col(rule.column) == rule.value).cast(daft.DataType.int64()).sum()
+            (col_name == rule.value).cast(daft.DataType.int64())
         )
[…]
