### 2024-05-17 canimus/cuallee#229 @FlorianK13
I also have problems running the check for `daft` where a template in the documentation could help a lot.
