### 2024-05-17 canimus/cuallee#229 @canimus
Thanks for the heads up @FlorianK13 . We purposely left out the Daft implementation in the paper because it has less than a year of use, though it was a commendable community contribution. We focused on six frameworks in our paper because I have professional experience with these and can confidently discuss the use of `cuallee` at scale, as well as its robustness and resilience within these frameworks.

For the JOSS submission, I believe covering the six highlighted frameworks—PySpark, Snowpark, Pandas, Polars, DuckDB, and BigQuery—already represents a significant effort. None of these implementations were trivial, and we only included those we considered mature and well-tested.

Given this context, I am open to addressing this issue but feel it is outside the scope of our current paper submission.

Agree?
