### 2024-05-17 canimus/cuallee#228 @canimus
Thank you @FlorianK13 we will include an additional section that address this request. Would you advise that we can then dismiss the word count limit of the paper?
