### 2024-05-17 canimus/cuallee#228 @FlorianK13
I don't know, maybe you can ask the editor in the Review Issue from JOSS?
