### 2024-05-18 93e7f24
Configuration for all tests with the expected libraries
