+++ cuallee/__init__.py
@@ -364,15 +364,16 @@ class Check:
         Rule("are_complete", column, "N/A", CheckDataType.AGNOSTIC, pct) >> self._rule
         return self
 
-    def is_unique(self, column: str, pct: float = 1.0):
+    def is_unique(self, column: str, pct: float = 1.0, approximate: bool = False):
         """
         Validation for unique values in column
         
         Args:
             column (str): Column name in dataframe
             pct (float): The threshold percentage required to pass
+            approximate (bool): A flag to speed up computation using an approximation through maximum relative std. dev.
         """
-        Rule("is_unique", column, "N/A", CheckDataType.AGNOSTIC, pct) >> self._rule
+        Rule("is_unique", column, "N/A", CheckDataType.AGNOSTIC, pct, options={"approximate" : approximate}) >> self._rule
         return self
 
     def is_primary_key(self, column: str, pct: float = 1.0):
