+++ test/performance/cuallee/test_performance_cuallee.py
@@ -1,17 +1,22 @@
-from pyspark.sql import SparkSession
+from os import truncate
 from cuallee import Check, CheckLevel
+
+from pyspark.sql import SparkSession
+import pyspark.sql.functions as F
 from datetime import datetime
 
 spark = SparkSession.builder.config("spark.driver.memory", "16g").getOrCreate()
 
 check = Check(CheckLevel.WARNING, "CualleeBenchmark")
-df = spark.read.parquet("data/*.parquet")
+df = spark.read.parquet("/data/*.parquet")
 
-for i in range(1000):
-    check.is_greater_than("fare_amount", i)
+
+[check.is_complete(name) for name in df.columns]
+[check.is_unique(name) for name in df.columns]
+
 
 start = datetime.now()
-check.validate(df).show(n=1000, truncate=False)
+check.validate(df).show(n=int(len(df.columns)*2), truncate=False)
 end = datetime.now()
 elapsed = end - start
 print("START:", start)
