+++ test/performance/soda/Dockerfile
@@ -7,16 +7,13 @@ RUN apt install openjdk-17-jdk -y
 WORKDIR /usr/src
 ENV PIP_ROOT_USER_ACTION=ignore
 RUN pip install pyspark==3.5.1
-RUN pip install pydeequ==1.3.0
-
-# Required to prevent an warning message on initialization
-ENV SPARK_VERSION="3.3"
+RUN pip install -i https://pypi.cloud.soda.io soda-spark-df==1.4.10
 
 RUN mkdir -p /data
 RUN wget https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2024-01.parquet -P /data
 RUN wget https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2024-02.parquet -P /data
 
-RUN mkdir -p /libs
-RUN wget https://repo1.maven.org/maven2/com/amazon/deequ/deequ/2.0.7-spark-3.5/deequ-2.0.7-spark-3.5.jar -P /libs
+ENV SODA_KEY=
+ENV SODA_SECRET=
 
 RUN rm -rf /root/.cache/pip
