+++ pyproject.toml
@@ -67,6 +67,9 @@ pdf = [
 daft = [
   "getdaft == 0.2.19"
 ]
+all = [
+  "cuallee[dev,pyspark,pyspark_connect,snowpark,pandas,bigquery,duckdb,polars,test,dagster,cloud,pdf,daft]"
+]
 
 [tool.ruff.lint]
 # Skip unused variable rules (`F841`).
