### 2024-05-18 960ed64
Feature performance frameworks (#231)
* Configuration for all tests with the expected libraries

* Bumped version for approx unique option
