+++ test/performance/cuallee/launcher.sh
@@ -1,1 +1,2 @@
-docker run -it -v $PWD/temp/taxi:/usr/src/data --rm --name cuallee-perftest cuallee-tester python /usr/src/sparker2.py
+echo "Running Performance Tests for Framework: CUALLEE"
+docker run -it -v $PWD:/usr/src --rm --name cuallee-tester-cuallee --rm cuallee-tester-cuallee python test_performance_cuallee.py
