+++ test/performance/greatexpectations/Dockerfile
@@ -7,7 +7,7 @@ RUN apt install openjdk-17-jdk -y
 WORKDIR /usr/src
 ENV PIP_ROOT_USER_ACTION=ignore
 RUN pip install pyspark==3.5.1
-RUN pip install cuallee==0.10.3
+RUN pip install great_expectations==0.18.13
 
 RUN mkdir -p /data
 RUN wget https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2024-01.parquet -P /data
