### 2024-05-18 9ad3212
Feature docs frameworks (#234)
* Added version for documents

* Working example documentation
