+++ docs/usage.md
@@ -1,164 +1,0 @@
-# Usage
-
-Validating data sets is about creating a `Check` and adding rules into it.
-You can choose from different types: `numeric`, `date algebra`, `range of values`, `temporal`, and many others.
-
-A `Check` provides a declarative interface to build a comprehensive validation on a dataframe as shown below:
-
-## Libraries
-
-```python
-# Imports
-from pyspark.sql import SparkSession
-import pyspark.sql.types as T
-import pyspark.sql.functions as F
-
-# Cuallee
-from cuallee import Check, CheckLevel
-from cuallee.pyspark_validation import _field_type_filter
-```
-
-## Session
-
-```python
-
-# Spark Session
-spark = SparkSession.builder.getOrCreate()
-
-```
-
-## Check
-
-```python
-# Check 
-check = Check(CheckLevel.WARNING, "TaxiNYCheck")
-
-```
-
-
-## Data
-```python
-# !wget https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2024-01.parquet
-# !wget https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2024-02.parquet
-df = spark.read.parquet("*yellow*tripdata*.parquet")
-df = df.withColumn("date", F.to_date("tpep_pickup_datetime"))
-```
-
-
-## Checks
-```python
-# Confirm that all records are not null for all columns in dataframe
-[check.is_complete(name) for name in df.columns]
-
-# Verify taxi ride distance is positive
-[check.is_greater_than(name, 0) for name in _field_type_filter(df, T.NumericType)] 
-
-# Confirm that tips are not outliers
-[check.is_less_than(name, 1e4) for name in _field_type_filter(df, T.NumericType)] 
-
-# 70% of data is on weekdays
-check.is_on_weekday("date", .7)
-
-# Binary classification fields
-[check.has_entropy(name, 1.0, 0.5) for name in _field_type_filter(df, T.NumericType)] 
-
-# Percentage of big tips
-[check.is_between(name, (1000,2000)) for name in _field_type_filter(df, T.NumericType)]
-
-# Confirm 22 years of data
-check.is_between("date", ("2000-01-01", "2022-12-31"))
-```
-
-
-## Validation
-```python
-check.validate(df).show()
-
-```
-
-If you want to display full results use:
-
-`check.validate(df).show(n=check.sum, truncate=False)`
-
-
-## Code
-
-??? example "Check"
-
-   
-
-    ``` python
-    # Imports
-    from pyspark.sql import SparkSession
-    import pyspark.sql.types as T
-    import pyspark.sql.functions as F
-
-    # Cuallee
-    from cuallee import Check, CheckLevel
-    from cuallee.pyspark_validation import _field_type_filter
-
-    # Session
-    spark = SparkSession.builder.getOrCreate()
-
-    # Check
-    check = Check(CheckLevel.WARNING, "TaxiNYCheck")
-
-    # !wget https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2024-01.parquet
-    # !wget https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2024-02.parquet
-    df = spark.read.parquet("*yellow*tripdata*.parquet")
-    df = df.withColumn("date", F.to_date("tpep_pickup_datetime"))
-
-    # Confirm that all records are not null for all columns in dataframe
-    [check.is_complete(name) for name in df.columns]
-
-    # Verify taxi ride distance is positive
-    [check.is_greater_than(name, 0) for name in _field_type_filter(df, T.NumericType)] 
-
[…]
