+++ docs/index.md
@@ -5,7 +5,7 @@ Welcome to `cuallee` and thanks for using this amazing framework. None of this w
 This `pure-python` implementation of unit tests for your data, will help you define validations for your data using `3` concepts described below:
 
 
-### Entities
+## Entities
 
 To better understand `cuallee` you will need to get familiar with the following 3 concepts: `Check`, `Rule` and `ComputeInstruction`. 
 
@@ -13,26 +13,39 @@ To better understand `cuallee` you will need to get familiar with the following 
 | ----------- | :------------------------------------: | ----------- |
 | `Check` | :material-check-all:   | Use it to define a group of validations on a `dataframe` and report them as `WARNING` or `ERROR`. You can chain as many rules into a `check`, internally `cuallee` will make sure the same rule is not executed twice. |
 | `Rule` | :material-check: | A `rule` represents the _predicate_ you want to test on a `single` or `multiple` columns in a dataframe. A rule as a 4 attributes `method`: name of the predicate, `column`: the column in the dataframe, `value`: the value to compare against and `coverage`: the percentage of positive predicate necessary to set the status of the check to `PASS`.  |
-| `ComputeInstruction` | :material-cog: | Are the implementation specific representations of the `predicates` in the `rule`. Because `cuallee` is a dataframe agnostic data quality framework, the implementation of the rules, rely in the creation of compute instructions passed to the specific dataframe of choice, including the following dataframe options: `pandas`, `pyspark` and `snowpark` |
+| `Control` | :material-gamepad: | Currently only supported in `pyspark` it allows to run pre-fabricated checks that target all columns in a dataframe. Controls help a user to run validation faster, without having to add individual rules for each column on a dataframe. For example `Control.completeness` creates a Check, and then adds `is_complete` rules to each column of the dataframe |
 
 In principle, the only interface you need to be familiar with is the `Check` as it is through this `object` that you can chain your validations and then directly through the `validate` method, execute validations on any `DataFrame`.
 
-## Process Flow
+## Process
+`cuallee` is designed to rapidly develop quality rules for your data.<br/>The process consists in the following :material-numeric-4-circle-outline: steps:
 ``` mermaid
 graph LR
-  U((start)) --> A;
-  A[Check] -.-> B(is_complete);
-  A[Check] -.-> C(is_between);
-  A[Check] -.-> D(is_on_weekday);
-  B --> E{all rules?};
-  C --> E{all rules?};
-  D --> E{all rules?};
-  E --> F[/read dataframe/];
-  A -.-> G{want results?};
-  F --> G;
-  G --> H(validate);
-  H --> I([get results])
-  I --> K((end))
+  
+  A[Create Check] -.-> B[Add Rules];
+  B[Add Rules] -.-> C[Load Data];
+  C[Load Data] -.-> D[Validate DataFrame];
 
 ```
 
+1. __Create Check__: Consists in calling the `check = Check()` constructor to initiate a validation
+2. __Add_Rules__: consists in calling any of the rules available in the check, providing a `column` and when required a expected `value`. For example: `check.is_unique(id)`
+3. __Load Data__: Is the process of creating a dataframe normally by loading data from your data sources, by convention in the docs we refer to dataframes as `df`. <br/>For example in `pandas` :material-arrow-right-bold-circle: `df = pd.read_csv("data.csv")`
+4. __Validate DataFrame__: Consists in running `check.validate(df)`
+
+## History
+
+Why did we create a software that offers functionality similar to `pydeequ`, `great-expectations`, and `soda-core`? While these tools are inspiring and have lowered the barrier for better data quality in the complex ecosystem of data platforms, they are primarily geared towards enterprise, collaborative, and licensed-supported software.
+
[…]
