+++ docs/dependencies.md
@@ -1,7 +1,25 @@
 # Dependencies
 
-It relies has only `2` dependencies that are automatically installed with the package:
+It relies has only :material-numeric-2-circle-outline: dependencies that are automatically installed with the package:
 
 - `toolz` for functional programming adoption
 - `requests` for downloading ISO standard required by `check.iso` interface.
-- `pandas` as the default validation result data abstraction
+
+## Metrics
+
+As a lightweight and near-zero-dependency framework `cuallee` excels by its simplicity.
+
+- `4.5k` lines of code :material-diamond-stone:
+- `<10` on code complexity for all validations :material-lightning-bolt:
+- `0` lines of code duplicated :material-gold:
+
+| Module      | Lines   | Complexity  |
+| ----------- | ------- | ----------- |
+| `pyspark`   |   `773`| `7` |
+| `pandas`   |   `343`| `0` |
+| `polars`   |   `450`| `2` |
+| `duckdb`   |   `232`| `4` |
+| `snowpark`   |   `817`| `3` |
+| `bigquery`   |   `275`| `4` |
+
+Reference: [scc](https://github.com/boyter/scc)
