### 2024-05-18 canimus/cuallee#229 @canimus
Reference:
- [duckdb](https://canimus.github.io/cuallee/duckdb/)
- [daft](https://canimus.github.io/cuallee/daft/)
- [catalogue](https://canimus.github.io/cuallee/catalogue/)

Fixed by #234
