### 2024-05-19 2552ec8
Addition of references for other frameworks
