+++ paper/paper.bib
@@ -75,3 +75,49 @@ keywords = {data annotation, data quality, data quality metrics, data utility, d
 location = {Corfu, Greece},
 series = {PETRA '22}
 }
+
+@inproceedings{10.1145/2723372.2742797,
+author = {Armbrust, Michael and Xin, Reynold S. and Lian, Cheng and Huai, Yin and Liu, Davies and Bradley, Joseph K. and Meng, Xiangrui and Kaftan, Tomer and Franklin, Michael J. and Ghodsi, Ali and Zaharia, Matei},
+title = {Spark SQL: Relational Data Processing in Spark},
+year = {2015},
+isbn = {9781450327589},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/2723372.2742797},
+doi = {10.1145/2723372.2742797},
+abstract = {Spark SQL is a new module in Apache Spark that integrates relational processing with Spark's functional programming API. Built on our experience with Shark, Spark SQL lets Spark programmers leverage the benefits of relational processing (e.g. declarative queries and optimized storage), and lets SQL users call complex analytics libraries in Spark (e.g. machine learning). Compared to previous systems, Spark SQL makes two main additions. First, it offers much tighter integration between relational and procedural processing, through a declarative DataFrame API that integrates with procedural Spark code. Second, it includes a highly extensible optimizer, Catalyst, built using features of the Scala programming language, that makes it easy to add composable rules, control code generation, and define extension points. Using Catalyst, we have built a variety of features (e.g. schema inference for JSON, machine learning types, and query federation to external databases) tailored for the complex needs of modern data analysis. We see Spark SQL as an evolution of both SQL-on-Spark and of Spark itself, offering richer APIs and optimizations while keeping the benefits of the Spark programming model.},
+booktitle = {Proceedings of the 2015 ACM SIGMOD International Conference on Management of Data},
+pages = {1383–1394},
+numpages = {12},
+keywords = {data warehouse, databases, hadoop, machine learning, spark},
+location = {Melbourne, Victoria, Australia},
+series = {SIGMOD '15}
+}
+
+@inproceedings{10.1145/2872427.2883029,
+author = {Pezoa, Felipe and Reutter, Juan and Suarez, Fernando and Ugarte, Martin and Vrgoč, Domagoj},
+year = {2016},
+month = {04},
+pages = {263-273},
+title = {Foundations of JSON Schema},
+doi = {10.1145/2872427.2883029}
+}
+
+@article{10.3389/fdata.2020.564115,
+  title={Toward Data Lakes as Central Building Blocks for Data Management and Analysis},
+  author={Dumke, André R. and Parchmann, Andreas and Schmid, Stefan and Hauswirth, Manfred},
+  journal={Frontiers in Big Data},
+  volume={3},
+  pages={564115},
+  year={2020},
+  publisher={Frontiers},
+  doi={10.3389/fdata.2020.564115}
+}
+
+@misc{oreilly2023technology,
+  title = {Technology Trends for 2023},
+  author = {{O'Reilly Media}},
+  year = {2023},
+  url = {https://www.oreilly.com/radar/technology-trends-for-2023/},
+  note = {Accessed: 2024-05-18}
+}
