### 2024-05-19 953b888
Added performance benchmark reference and versions
