+++ paper/paper.bib
@@ -121,3 +121,19 @@ doi = {10.1145/2872427.2883029}
   url = {https://www.oreilly.com/radar/technology-trends-for-2023/},
   note = {Accessed: 2024-05-18}
 }
+
+@misc{cuallee_performance_tests,
+  author       = {Herminio Vazquez},
+  title        = {cuallee: Performance Tests},
+  year         = {2024},
+  url          = {https://github.com/canimus/cuallee/tree/main/test/performance},
+  note         = {Accessed: 2024-05-19}
+}
+
+@misc{nyc_tlc_trip_record_data,
+  author       = {{New York City Taxi and Limousine Commission}},
+  title        = {TLC Trip Record Data},
+  year         = {2024},
+  url          = {https://www.nyc.gov/site/tlc/about/tlc-trip-record-data.page},
+  note         = {Accessed: 2024-05-19}
+}
