### 2024-05-19 f0f8d7c
Added performance benchmark reference and versions (#236)
