### 2024-05-21 canimus/cuallee#229 @FlorianK13
Hi @canimus, I was only testing the frameworks listed in the README, that's why I created this issue. I agree in the fact that the integration of several frameworks to a large number of test cases is not trivial.

I also saw that you have added the supported frameworks with examples to the documentation. From my opinion, this makes `cuallee` much easier to use for new users.
