### 2024-05-21 canimus/cuallee#229 @canimus
@FlorianK13 thank you very much for your attention to detail. Very much appreciated it. And very happy and glad with your observations. In the end, as you said we aim to support a long living user community and clarity is paramount of that. Have a great rest of your week.
