### 2024-05-21 7bbc0fe
- fixed has entropy
