### 2024-05-23 canimus/cuallee#239 @FlorianK13
[JOSS Review] State of the field in `paper.md`
This is probably my last issue of this review, as I have finished reviewing both the functionality and the documentation page. Only one minor thing about the recent changes in the paper:

* Both `soda` and `great-expectations` have a core component that is open source and additional features / cloud stuff where you need to pay. Could you correct this in your paper where you write:
*On the other hand, great-expectations and soda are commercial options that require registration and issuing of keys for cloud reporting capabilities.*
* Could you add references to the two repos? [great-expectation](https://github.com/great-expectations/great_expectations) has a citation.cff file for that purpose

This issue is part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6684)
