+++ paper/paper.bib
@@ -137,3 +137,16 @@ doi = {10.1145/2872427.2883029}
   url          = {https://www.nyc.gov/site/tlc/about/tlc-trip-record-data.page},
   note         = {Accessed: 2024-05-19}
 }
+
+@software{Gong_Great_Expectations,
+author = {Gong, Abe and Campbell, James and {Great Expectations}},
+license = {Apache-2.0},
+title = {{Great Expectations}},
+url = {https://github.com/great-expectations/great_expectations}
+}
+
+@software{soda_core,
+license = {Apache-2.0},
+title = {Soda Core},
+url = {https://github.com/sodadata/soda-core}
+}
