### 2024-05-24 fd5a742
modify text alternatives
