### 2024-05-24 3762e84
Feature paper ref (#240)
* add ref for great expectation and soda

* add ref great expectation and soda in paper

* modify text alternatives

* newest version pdf
