### 2024-05-26 643ed7c
Update and adjust daft to version `0.2.23` & above (#238)
* - Updated `daft_validation`
- Updated related tests
- Updated `pyproject.toml`

* - fixed has entropy

* - made `are_complete` backward compatible
