+++ test/unit/daft/test_are_complete.py
@@ -32,6 +32,8 @@ def test_coverage(check: Check):
     check.are_complete(("id", "id2"), 0.75)
     df = daft.from_pydict({"id": [10, None], "id2": [300, 500]})
     result = check.validate(df)
+
+    col_pass_rate = daft.col("pass_rate")
     assert (
-        result.select(daft.col("pass_rate").max() == 0.75).to_pandas().pass_rate.all()
+        result.agg(col_pass_rate.max()).select(col_pass_rate == 0.75).to_pandas().pass_rate.all()
     )
