### 2024-05-26 b73ef31
Amended pdf and corrected daft version in docs (#241)
