### 2024-05-26 e4c1242
Bump fpdf2 from 2.7.8 to 2.7.9 (#237)
Bumps [fpdf2](https://github.com/py-pdf/fpdf2) from 2.7.8 to 2.7.9.
- [Release notes](https://github.com/py-pdf/fpdf2/releases)
- [Changelog](https://github.com/py-pdf/fpdf2/blob/master/CHANGELOG.md)
- [Commits](https://github.com/py-pdf/fpdf2/compare/2.7.8...2.7.9)

---
updated-dependencies:
- dependency-name: fpdf2
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
