### 2024-05-27 a73b638
Bump duckdb from 0.10.2 to 0.10.3
Bumps [duckdb](https://github.com/duckdb/duckdb) from 0.10.2 to 0.10.3.
- [Release notes](https://github.com/duckdb/duckdb/releases)
- [Changelog](https://github.com/duckdb/duckdb/blob/main/tools/release-pip.py)
- [Commits](https://github.com/duckdb/duckdb/compare/v0.10.2...v0.10.3)

---
updated-dependencies:
- dependency-name: duckdb
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
