+++ pyproject.toml
@@ -45,7 +45,7 @@ bigquery =[
   "pyarrow >= 11.0.0"
 ]
 duckdb = [
-  "duckdb==0.10.2"
+  "duckdb==0.10.3"
 ]
 polars = [
   "polars>=0.19.6"
