### 2024-05-28 #6684 @FlorianK13
@jbytecode From my side the review is finished. The package `cuallee` from @canimus offers a large variety of test cases for different data frameworks in python. The implementation of the software, the software tests, and the documentation page all look very good. A clear recommendation from my side to publish the paper in JOSS.
