### 2024-05-29 canimus/cuallee#245 @devarops
[JOSS REVIEW] Great Expectations and Soda Core references are not rendering properly
We see that in the article, we are trying to cite Soda Core and Great Expectations:

https://github.com/canimus/cuallee/blob/e4c12429c5f440d3b2f4a64417ef7cc2737e0842/paper/paper.md?plain=1#L58

but it is not rendering properly:

![image](https://github.com/canimus/cuallee/assets/3461018/74d56af3-f65c-40a0-bf44-146df67539d8)

Maybe the `@software` option is not working.

https://github.com/canimus/cuallee/blob/e4c12429c5f440d3b2f4a64417ef7cc2737e0842/paper/paper.bib#L141-L152

We are confident that you can find a solution. Could you please consider using the `@misc` option or any other alternative that you think might work?

---

This issue is part of a [JOSS REVIEW](https://github.com/openjournals/joss-reviews/issues/6684)
