### 2024-05-30 canimus/cuallee#245 @vestalisvirginis
Thank you for your comment. But I don't understand what do you mean by 'it is not rendering properly' ? When running the pipeline to generate the manuscript, here is what we obtained:
- reference in the text:
- in the bibliography:

(See also paper.pdf in the paper directory in the repository)

which is different to the screen shot. Could it be that the paper hasn't been updated to the last version on your side?
