### 2024-05-30 e7c1e82
Added is_empty rule
