### 2024-06-03 73e4996
Bump getdaft from 0.2.24 to 0.2.25
Bumps [getdaft](https://github.com/Eventual-Inc/Daft) from 0.2.24 to 0.2.25.
- [Release notes](https://github.com/Eventual-Inc/Daft/releases)
- [Commits](https://github.com/Eventual-Inc/Daft/compare/v0.2.24...v0.2.25)

---
updated-dependencies:
- dependency-name: getdaft
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
