+++ pyproject.toml
@@ -65,7 +65,7 @@ pdf = [
   "fpdf2==2.7.9"
 ]
 daft = [
-  "getdaft == 0.2.24"
+  "getdaft == 0.2.25"
 ]
 all = [
   "cuallee[dev,pyspark,pyspark_connect,snowpark,pandas,bigquery,duckdb,polars,test,dagster,cloud,pdf,daft]"
