### 2024-06-04 canimus/cuallee#251 @devarops
[JOSS REVIEW] Run performance benchmark on CI pipeline
We were able to reproduce the performance benchmark manually. 🎉
However, seeing these results in the CI pipeline would be ideal. 🤖

Please add a job to the CI workflow on GitHub Actions to reproduce the performance benchmark.

---

This is the last issue of the review:
- https://github.com/openjournals/joss-reviews/issues/6684
