### 2024-06-10 a9441fb
Bump dagster from 1.7.6 to 1.7.9
Bumps [dagster](https://github.com/dagster-io/dagster) from 1.7.6 to 1.7.9.
- [Release notes](https://github.com/dagster-io/dagster/releases)
- [Changelog](https://github.com/dagster-io/dagster/blob/master/CHANGES.md)
- [Commits](https://github.com/dagster-io/dagster/compare/1.7.6...1.7.9)

---
updated-dependencies:
- dependency-name: dagster
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
