### 2024-06-18 b4eeb3e
Added is_empty rule (#246)
