### 2024-06-18 c2713a6
Fix all is-empty test cases
