+++ cuallee/__init__.py
@@ -58,12 +58,14 @@ except (ModuleNotFoundError, ImportError):
 
 class CheckLevel(enum.Enum):
     """Level of verifications in cuallee"""
+
     WARNING = 0
     ERROR = 1
 
 
 class CheckDataType(enum.Enum):
     """Accepted data types in checks"""
+
     AGNOSTIC = 0
     NUMERIC = 1
     STRING = 2
@@ -74,6 +76,7 @@ class CheckDataType(enum.Enum):
 
 class CheckStatus(enum.Enum):
     """Validation result criteria"""
+
     PASS = "PASS"
     FAIL = "FAIL"
     NO_RUN = "NO_RUN"
@@ -82,6 +85,7 @@ class CheckStatus(enum.Enum):
 @dataclass
 class Rule:
     """Predicate definition holder"""
+
     method: str
     column: Union[str, List[str], Tuple[str, str]]
     value: Optional[Any]
@@ -189,6 +193,7 @@ class Rule:
 
 class ComputeEngine(Protocol):
     """An interface for validatiosn to adhere to"""
+
     def compute(self, rules: Dict[str, Rule]) -> bool:
         """Returns compute instructions for each rule"""
 
@@ -216,9 +221,9 @@ class Check:
             level (CheckLevel): [0-1] value to describe if its a WARNING or ERROR check
             name (str): Normally the name of the dataset being verified, or a name for this check
             execution_date (date): An automatically generated timestamp of the check in UTC
-            table_name (str): When using databases matches the table name of the source 
+            table_name (str): When using databases matches the table name of the source
             session (Session): When operating in Session enabled environments like Databricks or Snowflake
-        
+
         """
         self._rule: Dict[str, Rule] = {}
         # TODO: Should be a compute engine protocol
@@ -271,7 +276,7 @@ class Check:
     def _remove_rule_generic(self, key: str):
         """
         Remove a key from rules and compute dictionaries
-        
+
         Args:
             key (str): the blake2s key of the rule
         """
@@ -281,7 +286,7 @@ class Check:
     def add_rule(self, method: str, *arg):
         """
         Add a new rule to the Check class.
-        
+
         Args:
             method (str): Check name
             arg (list): Parameters of the check
@@ -291,7 +296,7 @@ class Check:
     def delete_rule_by_key(self, keys: Union[str, List[str]]):
         """
         Delete rules from check based on keys.
-        
+
         Args:
             keys (List[str]): a single or list of keys to remove from the check
         """
@@ -308,7 +313,7 @@ class Check:
     ):
         """
         Delete rule based on method(s) or column name(s) or coverage value(s).
-        
+
         Args:
             rule_attribute (str): Finds a rule with by: method, column or coverage
             values (List[str]): Deletes a rule that matches the rule_attribute equal to the value in this parameter
@@ -326,13 +331,13 @@ class Check:
 
     def adjust_rule_coverage(self, rule_index: int, rule_coverage: float):
         """
-        Adjust the ratio predicate/rows for a rule. 
+        Adjust the ratio predicate/rows for a rule.
         It is intended to lower or increase tolerance without having to rewrite the entire check
 
         Args:
             rule_index (int): The position of the rule in the check list
             rule_coverage (float): New value between [0..1] for tolerance
-        
+
         """
         target_rule = self.rules[rule_index]
         old_key = target_rule.key
@@ -368,7 +373,7 @@ class Check:
     def are_complete(self, column: Union[List[str], Tuple[str, str]], pct: float = 1.0):
         """
         Validation for non-null values in a group of columns
-        
+
         Args:
             column (List[str]): A tuple or list of column names in dataframe
             pct (float): The threshold percentage required to pass
@@ -379,19 +384,29 @@ class Check:
[…]
