### 2024-06-18 29b4568
Bump getdaft from 0.2.24 to 0.2.27 (#253)
Bumps [getdaft](https://github.com/Eventual-Inc/Daft) from 0.2.24 to 0.2.27.
- [Release notes](https://github.com/Eventual-Inc/Daft/releases)
- [Commits](https://github.com/Eventual-Inc/Daft/compare/v0.2.24...v0.2.27)

---
updated-dependencies:
- dependency-name: getdaft
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
