### 2024-06-18 86ffad4
Bump duckdb from 0.10.2 to 1.0.0
Bumps [duckdb](https://github.com/duckdb/duckdb) from 0.10.2 to 1.0.0.
- [Release notes](https://github.com/duckdb/duckdb/releases)
- [Changelog](https://github.com/duckdb/duckdb/blob/main/tools/release-pip.py)
- [Commits](https://github.com/duckdb/duckdb/compare/v0.10.2...v1.0.0)

---
updated-dependencies:
- dependency-name: duckdb
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
