+++ pyproject.toml
@@ -45,7 +45,7 @@ bigquery =[
   "pyarrow >= 11.0.0"
 ]
 duckdb = [
-  "duckdb==0.10.2"
+  "duckdb==1.0.0"
 ]
 polars = [
   "polars>=0.19.6"
