### 2024-06-18 canimus/cuallee#251 @devarops
Hi @canimus,

Thank you for your detailed response. I appreciate the clarification regarding the performance evaluation.

The performance evaluation you provided is indeed manually reproducible with the Dockerfile and shell script included in the repository. As such, it is not a blocker issue. My suggestion was merely for consideration to enhance completeness.

Given your explanation, I am happy to recommend the publication of the paper in its current state.

Thank you for your efforts and dedication.

Cheers,
Evaristo
