### 2024-06-18 #6684 @devarops
@jbytecode: I am pleased to recommend the article "cuallee: A python package for data quality checks across multiple DataFrame APIs" for publication in JOSS. After a thorough review, the article significantly contributes to the journal. The software is original, high-quality, and represents a substantial scholarly effort. This article and the software will significantly interest the readers of JOSS, and I recommend it for publication.
