+++ paper/paper.bib
@@ -78,7 +78,7 @@ series = {PETRA '22}
 
 @inproceedings{10.1145/2723372.2742797,
 author = {Armbrust, Michael and Xin, Reynold S. and Lian, Cheng and Huai, Yin and Liu, Davies and Bradley, Joseph K. and Meng, Xiangrui and Kaftan, Tomer and Franklin, Michael J. and Ghodsi, Ali and Zaharia, Matei},
-title = {Spark SQL: Relational Data Processing in Spark},
+title = {Spark SQL: Relational Data Processing in {S}park},
 year = {2015},
 isbn = {9781450327589},
 publisher = {Association for Computing Machinery},
