### 2024-06-20 #6684 @jbytecode
@canimus - Meanwhile we are proof-reading the submission, you can go on with the archiving stuff. Please create a tagged release in the software repository. Using this tagged release, create a Zenodo archive. In this archive, the author list, their ORCIDs, the title of the manuscript should match. After creating the archive, please report here the tagged release, e.g., `v1.2.3`, and the DOI (and URL) of the Zenodo archive. Thank you in advance.
