### 2024-06-21 4938c53
Added fixes and release notes
