+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
 name = cuallee
-version = 0.10.4
+version = 0.10.5
 [options]
 packages = find:
