### 2024-06-21 d73cff4
Added fixes and release notes (#259)
