### 2024-06-21 15fc20e
Fix numpy issues
