### 2024-06-21 39fff4e
Fix numpy issues (#260)
