### 2024-06-21 9686ac9
Removal of the BQ assertions temporarily
