### 2024-06-21 f879dd2
Removal of the BQ assertions temporarily (#261)
