### 2024-06-21 92ec2fa
Added daft validations (#262)
