+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
 name = cuallee
-version = 0.10.5
+version = 0.11.0
 [options]
 packages = find:
