### 2024-06-21 #6684 @canimus
@jbytecode
Version: v0.11.0
URL: https://zenodo.org/records/12206787
DOI: 10.5281/zenodo.12206787
