### 2024-06-21 #6684 @jbytecode
@canimus - The metadata (the title, the list of authors, ORCIDs of authors, etc.) of the Zenodo archive and those of the submission should match. Could you please fix it and ping me again? Thank you in advance.
