### 2024-06-21 #6684 @canimus
Hi @jbytecode I decided to remove the DOI for that reference. Even when is published here:
https://www.frontiersin.org/articles/10.3389/fdata.2022.945720/full
Searching for that doi in the https://doi.org directly throws an error.
