### 2024-06-21 #6684 @jbytecode
It seems `10.3389/fdata.2022.945720` seems a valid DOI for the submission https://www.frontiersin.org/articles/10.3389/fdata.2022.945720/full
