### 2024-06-21 #6684 @canimus
@jbytecode the problem is that when passing through the bot verification it qualifies it as INVALID
