### 2024-06-21 #6684 @jbytecode
@canimus - `10.3389/fdata.2020.564115` was not passing, but `10.3389/fdata.2022.945720` is resolved. Am I wrong? You can add the same bibtex entry with the latest DOI, right?
