+++ paper/paper.bib
@@ -103,6 +103,16 @@ title = {Foundations of JSON Schema},
 doi = {10.1145/2872427.2883029}
 }
 
+@article{10.3389/fdata.2022.945720,
+  title={Toward Data Lakes as Central Building Blocks for Data Management and Analysis},
+  author={Dumke, André R. and Parchmann, Andreas and Schmid, Stefan and Hauswirth, Manfred},
+  journal={Frontiers in Big Data},
+  volume={3},
+  pages={564115},
+  year={2020},
+  publisher={Frontiers},
+  doi={10.3389/fdata.2022.945720}
+}
 
 @misc{oreilly2023technology,
   title = {Technology Trends for 2023},
