### 2024-06-21 #6684 @jbytecode
@canimus - It seems we resolved the DOI issue. Now I am generating the latest pdf and I'll have a final proof-reading.
