### 2024-06-21 #6684 @canimus
@jbytecode I downloaded the latest PDF version, and I am confident that is now good to go. It is free from typos, and readability score is good. Thank you very much for all your support, assistance and diligence, truly appreciate it.
