### 2024-06-21 #6684 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
