### 2024-06-21 #6684 @jbytecode
@crvernon - `10.5040/9781350207318.00000004 may be a valid DOI for title: Great Expectations` refers to an other manuscript with the same exact title, the other bibtex entries haven't got DOIs. fyi.
