### 2024-06-23 #6684 @crvernon
:wave: @canimus - Thank you for taking such care to produce a really nice paper and product!  And great job editing @jbytecode!

The following is just one very minor thing that need to be fixed before moving forward:

- [ ] In your Zenodo archive, please capitalize the "p" in "python" to match the correct capitalization that is represented in the paper.  You just have to edit the title via the metadata; no need to make a new release.

Great work!
