### 2024-06-23 #6684 @canimus
@crvernon - Thank you for your feedback, truly appreciated. Both Virginie and myself put our hearth in it, and we are looking forward for its use in the community. Today we literally breakfast thinking in Bio checks as the next step in cuallee.

I have changed the name in the Zenodo archive as suggested.
